% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{abbrev}
\alias{abbrev}
\title{Create vector of Initials}
\usage{
abbrev(x, exceptions = NULL)
}
\arguments{
\item{x}{vector of strings to be abbreviated.}

\item{exceptions}{list with two names vectors:  \code{old}, a vector abbreviations to 
be replaced and  \code{new}, a vector of replacment values.}
}
\value{
a character vector of \code{length(x)}.
}
\description{
\code{abbrev} returns a character vector of each words first capital letters for each element of \code{x}.
}
\details{
This function returns a same-length character vector
that abbrevs an initial character vector, \code{x}.  Abbreviation returns the first
capital letter of any words in each element of \code{x}.  Users may additionally pass
\code{abbrev} an optional list of exceptions that overides the default abbreviations.
The list of exceptions requires a vector of "old" values to be replaced by "new" values
}
\examples{
x<-c("KIPP Ascend Middle School", "KIPP Ascend Primary School", 
     "KIPP Create College Prep", 
     "KIPP Bloom College Prep" ,
     "KIPP One Academy")

abbrev(x)

altnames<-list(old=c("KAPS", "KBCP", "KOA"), 
                 new=c("KAP", "Bloom", "One"))

abbrev(x, exceptions=altnames)
}

