% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_data.R
\docType{data}
\name{act_df}
\alias{act_df}
\title{ACT linking data frame}
\format{\describe{
\item{school}{I think this was put into the data to allow for rbinding
with a different data set.  Can be ignored.}
\item{cohort}{similar to school, can be ignored.}
\item{grade}{grade level season for the observation}
\item{rit}{RIT equivalency}
\item{act}{ACT sub-test equivalency}
\item{subject}{Mathematics or Reading}
}}
\source{
NWEA linking study
}
\usage{
act_df
}
\description{
RIT/ACT equivalencies derived from NWEA ACT linking study
}
\keyword{datasets}
