% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goals_and_projections.R
\name{add_accelerated_growth}
\alias{add_accelerated_growth}
\title{Add an accelerated growth object, including projections and 
simulations, to a mapvizieR object}
\usage{
add_accelerated_growth(mapvizier_object, goal_function = goal_kipp_tiered,
  goal_function_args = list(iterations = 1), update_growth_df = FALSE)
}
\arguments{
\item{mapvizier_object}{a \code{\link{mapvizieR}} object.}

\item{goal_function}{a function that returns a list containing a a data 
frame named \code{goals}, a character vector of columns used to join 
accelerated goals to \code{growth_df}, and \code{slot_name} single element
character vector used to name the slot in the \code{goals} element of 
a \code{mapvizieR} object.}

\item{goal_function_args}{arguments passed to \code{goal_function}}

\item{update_growth_df}{if \code{TRUE} accelerated growth and met accelerated
growth columns are added/updated in the \code{growth_df} of a \code{mapvizieR}
object}
}
\value{
a \code{\link{mapvizieR}} object.
}
\description{
\code{add_accelerated_growth} is a constructor function that adds a 
"goals" object (a list with a \code{goals} data frame, a \code{join_by_fields}
character vector, and \code{slot_name} single element character vector) to
a \code{\link{mapvizieR}} object. The goals object is added to a \code{goals}
slot in the \code{mapvizieR} object. The goals themselves, as well as any 
projections or simulations, are created by a "goals function" (see \code{\link{goal_kipp_tiered}}
for an example) that is passed as the \code{goal_function argument}; 
arguments to the \code{goal_function} are passed via the \code{goal_function_args} 
argument. Note well that the \code{goal_function} must (i) return a list with 
three elements (the goals data frame, the join_by_fields character vector,
and the slot_name) and (ii) the goals data frame must have at least fields named
\code{accel_growth} and \code{met_accel_growth}. If the \code{updated_growth_df}
is TRUE then the goals data frame is \code{inner_join}ed  with the 
\code{growth_df} using the \code{join_by_fields}, accelerated growth columns are added or updated, and any 
duplicate columns from the join are cleaned up (original columns from 
the \code{growth_df} are maintened). Obviouslly, the goals function should
return a one to one match on any first iterations.
}
\examples{
\dontrun{
data(ex_CombinedAssessmentResults)
data(ex_CombinedStudentsBySchool)

cdf_mv <- mapvizieR(
 ex_CombinedAssessmentResults, 
 ex_CombinedStudentsBySchool
)
                    
new_mv <- add_accelerated_growth(
 cdf_mv,
 goal_function = goal_kipp_tiered, 
 goal_function_args = list(iterations=1),
 update_growth_df = FALSE
)
str(new_mv)                                
}
}

