% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{adhoc_psuedo_quartile}
\alias{adhoc_psuedo_quartile}
\title{ad-hoc psuedo quartile}
\usage{
adhoc_psuedo_quartile(x, breaks)
}
\arguments{
\item{breaks}{vector of breaks}

\item{vector}{to compare (presumably percentile ranks)}
}
\value{
vector of 'quartiles'
}
\description{
can be used to get becca plot to return predicted performance 
bands, instead of quartile bands
}

