% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_cohort_cgp_hist.R
\name{alt_cohort_cgp_hist_plot}
\alias{alt_cohort_cgp_hist_plot}
\title{Shows a cohort's progress over time.  Similar to cohort_cgp_hist,
but uses the 2015 school grade level attainment/status norms.}
\usage{
alt_cohort_cgp_hist_plot(mapvizieR_obj, studentids, measurementscale,
  match_method = "no matching", first_and_spring_only = TRUE,
  entry_grade_seasons = c(-0.8, 4.2), primary_cohort_only = TRUE,
  small_n_cutoff = 0.5, no_labs = FALSE, labels = "RIT")
}
\arguments{
\item{mapvizieR_obj}{conforming mapvizieR obj}

\item{studentids}{vector of studentids}

\item{measurementscale}{target subject}

\item{match_method}{do we limit to matched students, and if so, how?
no matching = any student record in the studentids.
UNIMPLEMENTED METHODS / TODO
strict = only kids who appear in all terms
strict after imputation = impute first, then use stritc method
back one = look back one test term, and only include kids who can be matched}

\item{first_and_spring_only}{show all terms, or only entry & spring?
default is TRUE.}

\item{entry_grade_seasons}{which grade_level_seasons are entry grades?}

\item{primary_cohort_only}{will determine the most frequent cohort and limit to 
students in that cohort.  designed to handle discrepancies in grade/cohort
pattern caused by previous holdovers.  default is TRUE.}

\item{small_n_cutoff}{any cohort below this percent will get filtered out.
default is 0.5, eg cohorts under 0.5 of max size will get dropped.}

\item{no_labs}{if TRUE, will not label x or y axis}

\item{labels}{c('RIT', 'NPR').  'RIT' is default.}
}
\value{
a ggplot object
}
\description{
Shows a cohort's progress over time.  Similar to cohort_cgp_hist,
but uses the 2015 school grade level attainment/status norms.
}

