% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/amys_list.R
\name{amys_lists}
\alias{amys_lists}
\title{Students' names stacked by category, or a more informative bar graph.}
\usage{
amys_lists(mapvizieR_obj, studentids, measurementscale, start_fws,
  start_academic_year, end_fws, end_academic_year)
}
\arguments{
\item{mapvizieR_obj}{mapvizieR object}

\item{studentids}{target students}

\item{measurementscale}{target subject}

\item{start_fws}{starting season}

\item{start_academic_year}{starting academic year}

\item{end_fws}{ending season}

\item{end_academic_year}{ending academic year}
}
\description{
\code{amys_lists} implements bar charts that created by ranking and
stacking student names for each growth status bin (i.e., negative growth,
not typical, typical, and college ready growth).  The name is from KIPP Chicago
CAO Amy Pouba, who came up with the idea.  This visualizaiton is nice in that it provides
a quick overview of the distribution of growth statuses while simultaneously providing
student-level information (student name, end RIT score, and gorwth status).
}

