% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_window.R
\name{auto_growth_window}
\alias{auto_growth_window}
\title{Extract likely growth window}
\usage{
auto_growth_window(
  mapvizieR_obj,
  studentids,
  measurementscale,
  end_fws,
  end_academic_year,
  candidate_start_fws = c("Fall", "Spring"),
  candidate_year_offsets = c(0, -1),
  candidate_prefer = "Spring",
  window_tolerance = 0.5
)
}
\arguments{
\item{mapvizieR_obj}{a mapvizier object}

\item{studentids}{the studentids in question}

\item{measurementscale}{a NWEA MAP measurementscale}

\item{end_fws}{desired end of growth term (season)}

\item{end_academic_year}{desired end of growth term (year)}

\item{candidate_start_fws}{two seasons to pick from}

\item{candidate_year_offsets}{if prev spring, -1}

\item{candidate_prefer}{which one is the 'best' term?}

\item{window_tolerance}{revert to the other term if this one is below the tolerance}
}
\value{
a list with inferred start season and year
}
\description{
utility function to determine the best growth window from the data presented
}
