% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_df_prep.R
\name{build_growth_scaffolds}
\alias{build_growth_scaffolds}
\title{build_growth_scaffolds}
\usage{
build_growth_scaffolds(processed_cdf)
}
\arguments{
\item{processed_cdf}{conforming mapvizieR processed cdf}
}
\description{
a helper function for \code{generate_growth_df}.
finds all the student/season growth windows.
}
