% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/college_plots.R
\name{build_student_college_plot}
\alias{build_student_college_plot}
\title{Combines a template and a mapvizieR object to create a student goal plot
with college labels}
\usage{
build_student_college_plot(
  base_plot,
  mapvizieR_obj,
  studentid,
  measurementscale,
  labels_at_grade,
  localization = localize("Newark"),
  aspect_ratio = 1
)
}
\arguments{
\item{base_plot}{a height/weight template.}

\item{mapvizieR_obj}{conforming mapvizieR object}

\item{studentid}{target student}

\item{measurementscale}{target subject}

\item{labels_at_grade}{what grade level should the college labels 
print at?  Generally the students's most recent test grade_level_season
is desirable.}

\item{localization}{controls names/breakpoints for college labels and
ACT tiers.  See localization.R for more details}

\item{aspect_ratio}{college labels should print at the same
angle as the act ribbons.  if the viz is not square, this requires an
adjustment.  default value is 1, 0.5 would be a rectangle 2W for 1H.}
}
\value{
a ggplot object
}
\description{
Combines a template and a mapvizieR object to create a student goal plot
with college labels
}
