% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{build_year_in_district}
\alias{build_year_in_district}
\title{build_year_in_district}
\usage{
build_year_in_district(roster)
}
\arguments{
\item{roster}{a roster data frame. must contain studentid, map_year_academic,
and grade}
}
\value{
data frame with disambiguated termname and map_year_academic
}
\description{
tags the roster with the student's year in the district
}

