% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cgp_prep.R
\name{calc_cgp}
\alias{calc_cgp}
\title{calc_cgp}
\usage{
calc_cgp(measurementscale, grade, growth_window, baseline_avg_rit = NA,
  ending_avg_rit = NA, sch_growth_study = sch_growth_norms_2012,
  calc_for = c(1:99))
}
\arguments{
\item{measurementscale}{MAP subject}

\item{grade}{baseline/starting grad for the group of students}

\item{growth_window}{desired growth window for targets (fall/spring, spring/spring, fall/fall)}

\item{baseline_avg_rit}{the baseline mean rit for the group of students}

\item{ending_avg_rit}{the baseline mean rit for the group of students}

\item{sch_growth_study}{a school growth study to use.  default is sch_growth_norms_2012}

\item{calc_for}{vector of cgp targets to calculate for.}
}
\value{
a named list - targets, and results
}
\description{
calculates both cgp targets and cgp results.
}

