% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goals_and_projections.R
\name{calc_normed_student_growth}
\alias{calc_normed_student_growth}
\title{Calculate accelerated growth from norms using a target percentile}
\usage{
calc_normed_student_growth(percentile, typical_growth, sd_growth)
}
\arguments{
\item{percentile}{the target percentile must be between 0 and 1 or 0 and 100}

\item{typical_growth}{the student's expected growth}

\item{sd_growth}{the standard deviation of expected growth}
}
\value{
a numeric vector of accelerated growth
}
\description{
Calculate accelerated growth from norms using a target percentile
}
\examples{
calc_normed_student_growth(.75, 5, 2)
calc_normed_student_growth(75, 5, 2) 

}

