% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/growth_df_prep.R
\name{calc_rit_growth_metrics}
\alias{calc_rit_growth_metrics}
\title{Calculate growth metrics (RIT growth, meeting indicators, conditional
change in test percentile, growth index, student growth percentile).}
\usage{
calc_rit_growth_metrics(normed_df)
}
\arguments{
\item{normed_df}{a data frame that has matched growth windows for
each student/subject/season triplet.}
}
\value{
a data frame the same as growth_df, with additional calcs
}
\description{
a helper function  for \code{generate_growth_df}
which adds columns to a growth CDF for variuos growth statistics.
}

