% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{cdf_collapse_by_grade}
\alias{cdf_collapse_by_grade}
\title{cdf collapse by grade}
\usage{
cdf_collapse_by_grade(cdf)
}
\arguments{
\item{cdf}{}
}
\value{
data frame
}
\description{
strips out term and year info so that summary.mapvizieR_cdf will cleanly group students by grade_level_season, school
and measurementscale
}

