% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cdf_prep.R
\name{cdf_prep_fws}
\alias{cdf_prep_fws}
\title{cdf_prep_fws}
\usage{
cdf_prep_fws(cdf_long)
}
\arguments{
\item{cdf_long}{a map assessmentresults.csv file.  can be one term, or many terms
together in one file.}
}
\value{
a cdf with new term fields
}
\description{
\code{cdf_prep_fws} munges test term data on a CDF
}

