% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cdf_prep.R
\name{cdf_prep_names}
\alias{cdf_prep_names}
\title{cdf_prep_names}
\usage{
cdf_prep_names(cdf_long)
}
\arguments{
\item{cdf_long}{a map assessmentresults.csv file.  can be one term, or many terms
together in one file.}
}
\value{
a cdf with lowercase variable names
}
\description{
\code{cdf_prep_names} turns the CamelCase names of a cdf to lowercase.
}

