% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapvizieR_object.R
\name{cdf_roster_match}
\alias{cdf_roster_match}
\title{match assessment results with students by school roster.}
\usage{
cdf_roster_match(assessment_results, roster)
}
\arguments{
\item{assessment_results}{a cdf file that passes the checks in \code{\link{check_cdf_long}}}

\item{roster}{a roster that passes the checks in \code{\link{check_roster}}}
}
\value{
a merged data frame with \code{nrow(prepped_cdf)}
}
\description{
\code{cdf_roster_match} performs an inner join on a prepped, long cdf (Assesment Results)
a prepped long roster (i.e. StudentsBySchool).
}

