% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_detail_table.R
\name{cdf_stu_change}
\alias{cdf_stu_change}
\title{find the RIT and NPR change across a cdf}
\usage{
cdf_stu_change(cdf)
}
\arguments{
\item{cdf}{a processed/prepped CDF, limited to the students and time 
period desired}
}
\value{
a data frame with studentid, measurementscale, rit start/end,
rit change, npr start/end, npr change
}
\description{
looks at a CDF, takes first/last event by stu/subj, reports
change in RIT and NPR
}
