% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/college_plots.R
\name{centroid}
\alias{centroid}
\title{Finds the centroid of a triangle}
\usage{
centroid(x1, x2, x3, y1, y2, y3)
}
\arguments{
\item{x1}{first x coord}

\item{x2}{second x coord}

\item{x3}{third x coord}

\item{y1}{first y coord}

\item{y2}{second y coord}

\item{y3}{third y coord}
}
\value{
a numeric vector length 2 with the xy coord of the centroid
}
\description{
Finds the centroid of a triangle
}

