% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cgp_prep.R
\name{cgp_target_lookup}
\alias{cgp_target_lookup}
\title{cgp_target_lookup}
\usage{
cgp_target_lookup(measurementscale_in, grade_in, growth_window_in,
  baseline_avg_rit, sch_growth_study = sch_growth_norms_2012)
}
\arguments{
\item{baseline_avg_rit}{the baseline mean rit for the group of students}

\item{measurementscale}{MAP subject}

\item{grade}{baseline/starting grad for the group of students}

\item{growth_window}{desired growth window for targets (fall/spring, spring/spring, fall/fall)}
}
\description{
get the best matching row from a school growth study data frame.
}

