% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/cdf_check.R
\name{check_cdf_fws}
\alias{check_cdf_fws}
\title{check_cdf_fws}
\usage{
check_cdf_fws(prepped_cdf_long)
}
\arguments{
\item{prepped_cdf_long}{a map CDF file, generated either by prep_cdf_long,
or via processing done in your data warehouse}
}
\value{
boolean, does the cdf pass?
}
\description{
\code{check_cdf_fws} tests the fallwinterspring column of a prepped cdf to
see if it conforms
}

