% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf_check.R
\name{check_cdf_long}
\alias{check_cdf_long}
\title{check_cdf_long}
\usage{
check_cdf_long(prepped_cdf_long)
}
\arguments{
\item{prepped_cdf_long}{a map CDF file, generated either by prep_cdf_long,
or via processing done in your data warehouse}
}
\value{
a named list.  \code{$boolean} has true false result; \code{descriptive} 
has a more descriptive string describing what happened.
}
\description{
\code{check_cdf_long} a wrapper around a bunch of individual tests
that see if a CDF data frame conforms to mapvizieR conventions
}

