% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/cdf_check.R
\name{check_cdf_names}
\alias{check_cdf_names}
\title{check_cdf_names}
\usage{
check_cdf_names(prepped_cdf_long)
}
\arguments{
\item{prepped_cdf_long}{a map CDF file, generated either by prep_cdf_long,
or via processing done in your data warehouse}
}
\value{
boolean, does the cdf pass?
}
\description{
\code{check_cdf_names} tests a prepped cdf's names to see if they conform
}

