% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mapvizieR_object.R
\name{check_is_mapvizieR}
\alias{check_is_mapvizieR}
\title{check_is_mapvizieR}
\usage{
check_is_mapvizieR(x)
}
\arguments{
\item{x}{an object to test}
}
\value{
boolean
}
\description{
tests if the object is a mapvizieR object and throws a descriptive
error if not
}

