% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf_check.R
\name{check_processed_cdf}
\alias{check_processed_cdf}
\title{check_processed_cdf}
\usage{
check_processed_cdf(processed_cdf)
}
\arguments{
\item{processed_cdf}{output of mapvizieR.default}
}
\value{
a named list.  \code{$boolean} has true false result; \code{descriptive} 
has a more descriptive string describing what happened.
}
\description{
the mapvizieR takes a cdf + a roster and does some grade level lookup.
this function is a wrapper around some tests that make sure that
the output conforms to expectations
}

