% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/cdf_check.R
\name{check_processed_names}
\alias{check_processed_names}
\title{check_processed_names}
\usage{
check_processed_names(processed_cdf)
}
\arguments{
\item{processed_cdf}{output of mapvizieR.default}
}
\value{
boolean, does the processed cdf pass?
}
\description{
does the processed cdf have the right field names
}

