% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/roster_check.R
\name{check_roster}
\alias{check_roster}
\title{check_roster}
\usage{
check_roster(roster)
}
\arguments{
\item{roster}{a roster file, generated either by prep_roster,
or via processing done in your data warehouse}
}
\value{
a named list.  \code{$boolean} has true false result; \code{descriptive}
has a more descriptive string describing what happened.
}
\description{
\code{check_roster} a wrapper around a bunch of individual tests
that see if a roster data frame conforms to mapvizieR expectations
}

