% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/roster_check.R
\name{check_roster_names}
\alias{check_roster_names}
\title{check_roster_names}
\usage{
check_roster_names(roster)
}
\arguments{
\item{roster}{a roster file, generated either by prep_roster,
or via processing done in your data warehouse}
}
\value{
boolean
}
\description{
\code{check_roster_names} does the roster have a studentid field?
}

