% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/util.R
\name{clean_measurementscale}
\alias{clean_measurementscale}
\title{clean_measurementscale}
\usage{
clean_measurementscale(x)
}
\arguments{
\item{x}{a measurementscale}
}
\description{
add any logic about cleaning measurementscales to this function
}

