% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgp_prep.R
\name{cohort_expectation}
\alias{cohort_expectation}
\title{cohort_expectation}
\usage{
cohort_expectation(measurementscale, end_grade, growth_window,
  baseline_avg_rit, calc_for, norms = 2015)
}
\arguments{
\item{measurementscale}{a MAP subject}

\item{end_grade}{the ENDING grade level for the growth window.  ie, if this calculation
crosses school years, use the grade level for the END of the term, per the example on p. 7
of the 2012 school growth study}

\item{growth_window}{the growth window to calculate CGP over}

\item{baseline_avg_rit}{mean rit at the START of the growth window}

\item{calc_for}{what CGPs to calculate for?}

\item{norms}{which school growth study to use.  c(2012, 2015).  default is
2015.}
}
\description{
wrapper function to get cohort growth expectations for the lookup method
}
