% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cgp_prep.R
\name{cohort_expectation_generalization}
\alias{cohort_expectation_generalization}
\title{cohort_expectation_generalization}
\usage{
cohort_expectation_generalization(measurementscale_in, grade_in,
  growth_window_in, baseline_avg_rit)
}
\arguments{
\item{baseline_avg_rit}{the baseline mean rit for the group of students}

\item{measurementscale}{MAP subject}

\item{grade}{baseline/starting grad for the group of students}

\item{growth_window}{desired growth window for targets (fall/spring, spring/spring, fall/fall)}
}
\description{
get the best matching row from a school growth study data frame.
}

