% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cgp_prep.R
\name{cohort_expectation_via_generalization}
\alias{cohort_expectation_via_generalization}
\title{cohort_expectation_via_generalization}
\usage{
cohort_expectation_via_generalization(measurementscale_in, grade_in,
  growth_window_in, baseline_avg_rit, baseline_avg_npr, sch_growth_study,
  calc_for)
}
\arguments{
\item{baseline_avg_rit}{the baseline mean rit for the group of students}
}
\description{
expectation in NPR via generalization
}

