% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_longitudinal_plots.R
\name{cohort_longitudinal_npr_plot}
\alias{cohort_longitudinal_npr_plot}
\title{Longitudinal plot against NPR background}
\usage{
cohort_longitudinal_npr_plot(mapvizieR_obj, studentids, measurementscale,
  first_and_spring_only = TRUE, entry_grade_seasons = c(-0.8, 4.2),
  name_annotations = FALSE, student_alpha = 0.1, trace_lines = c(1, 10,
  20, 30, 40, 50, 60, 70, 80, 90, 99))
}
\arguments{
\item{mapvizieR_obj}{a conforming mapvizieR object}

\item{studentids}{a vector of studentids}

\item{measurementscale}{target subject}

\item{first_and_spring_only}{logical, should we include fall/winter scores
from non-entry grades?}

\item{entry_grade_seasons}{what grades are 'entry' grades for this school?}

\item{name_annotations}{should we include student names on the plot? 
default is FALSE.}

\item{student_alpha}{how much to alpha-out the student observations?}

\item{trace_lines}{what norms to show?}
}
\value{
a ggplot object
}
\description{
shows the progress of students, and cohort average, against
the NPR space.
}
