% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgp_prep.R
\name{cohort_mean_rit_to_npr}
\alias{cohort_mean_rit_to_npr}
\title{cohort_mean_rit_to_npr}
\usage{
cohort_mean_rit_to_npr(measurementscale, current_grade, season, RIT)
}
\arguments{
\item{measurementscale}{MAP subject}

\item{current_grade}{grade level}

\item{season}{fall winter spring}

\item{RIT}{mean rit score}
}
\value{
a integer vector length one
}
\description{
given a mean RIT score for a cohort, return the best
 match percentile rank.  (assumes the subject is grade/cohort, not a student.)
only supports 2015 norms.
}
