% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_status_trace_plot.R
\name{cohort_status_trace_plot}
\alias{cohort_status_trace_plot}
\title{Cohort Status Trace Plot}
\usage{
cohort_status_trace_plot(mapvizieR_obj, studentids, measurementscale,
  match_method = "no matching", first_and_spring_only = TRUE,
  entry_grade_seasons = c(-0.8, 4.2), collapse_schools = TRUE,
  retention_strategy = "collapse", small_n_cutoff = 0.5,
  plot_labels = "RIT")
}
\arguments{
\item{mapvizieR_obj}{conforming mapvizieR obj}

\item{studentids}{vector of studentids}

\item{measurementscale}{target subject}

\item{match_method}{do we limit to matched students, and if so, how?
no matching = any student record in the studentids.
UNIMPLEMENTED METHODS / TODO
strict = only kids who appear in all terms
strict after imputation = impute first, then use stritc method
back one = look back one test term, and only include kids who can be matched}

\item{first_and_spring_only}{show all terms, or only entry & spring?
default is TRUE.}

\item{entry_grade_seasons}{which grade_level_seasons are entry grades?}

\item{collapse_schools}{treats all students as part of the same 'school' for purposes of plotting, so that one trajectory is shown.
default is TRUE.  if FALSE will separate lines by school and show a lengend.}

\item{retention_strategy}{}

\item{plot_labels}{c('RIT', 'NPR').  'RIT' is default.}

\item{mapvizieR_obj}{}

\item{studentids}{}

\item{measurementscale}{}

\item{match_method}{}

\item{first_and_spring_only}{}

\item{entry_grade_seasons}{}

\item{collapse_schools}{}

\item{plot_labels}{}
}
\value{
a ggplot object

a ggplot object
}
\description{
Cohort Status Trace Plot
}

