% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/college_plots.R
\name{college_label_element}
\alias{college_label_element}
\title{abstract out the logic for writing college labels onto a template.}
\usage{
college_label_element(xy_lim_list, measurementscale, labels_at_grade = 6,
  localization = localize("Newark"), aspect_ratio = 1, label_size = 5)
}
\arguments{
\item{xy_lim_list}{student max/min x and y, from stu_RIT_hist_plot_elements}

\item{measurementscale}{target subject}

\item{labels_at_grade}{what grade level should the college labels
print at?  Generally the student or cohorts's most recent test
grade_level_season is desirable.}

\item{localization}{controls names/breakpoints for college labels and
ACT tiers.  See localization.R for more details}

\item{aspect_ratio}{college labels should print at the same
angle as the act ribbons.  if the viz is not square, this requires an
adjustment.  default value is 1, 0.5 would be a rectangle 2W for 1H.}

\item{label_size}{text size for the labels.  default is 5.}
}
\value{
a ggplot geom_text object with college labels
}
\description{
this was non-trivial, so it gets its own little function.
}

