% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mapvizieR_object.R
\name{create_mapvizier_object}
\alias{create_mapvizier_object}
\title{create_mapvizier_object}
\usage{
create_mapvizier_object(raw_cdf, raw_roster)
}
\arguments{
\item{raw_cdf}{a NWEA AssessmentResults.csv or CDF}

\item{raw_roster}{a NWEA students}
}
\description{
\code{create_mapvizier_object} is a workhorse workflow function that
calls a sequence of cdf and roster prep functions, given a raw cdf and raw roster
}

