% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cdf_prep.R
\name{dedupe_cdf}
\alias{dedupe_cdf}
\title{dedupe_cdf}
\usage{
dedupe_cdf(prepped_cdf, method = "NWEA")
}
\arguments{
\item{prepped_cdf}{conforming prepped cdf file.}

\item{method}{can choose between c('NWEA', 'high RIT', 'most recent').
Default is NWEA method.}
}
\value{
a data frame with one row per kid
}
\description{
\code{dedupe_cdf} makes sure that the cdf only contains one row student/subject/term
}

