% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cgp_prep.R
\name{determine_cgp_method}
\alias{determine_cgp_method}
\title{determine_cgp_method}
\usage{
determine_cgp_method(measurementscale_in, grade_in, growth_window_in,
  baseline_avg_rit, tolerance, sch_growth_study)
}
\arguments{
\item{baseline_avg_rit}{the baseline mean rit for the group of students}

\item{tolerance}{NWEA has published empirical lookup tables for growth.  these tables cover
the middle of the distribution, meaning that data for most cohorts can be found in the table.
but say that you have a cohort that is 10 rit points below the lowest mean RIT reported in the
tables?  do you just use the lowest entry?
mapvizieR has a generalization of the school growth study that translate the RIT changes to
percentile change, and then fits a general model to express cohort growth percentile
given a starting percentile and an ending percentile. rather than use lookup tables that we
know to be poor fits to the students}
}
\description{
given all the cgp parameters, determines how we'll look up growth scores.
}

