% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/growth_df_prep.R
\name{determine_growth_status}
\alias{determine_growth_status}
\title{determine growth status}
\usage{
determine_growth_status(df)
}
\arguments{
\item{df}{output of add accelerated growth}
}
\description{
last thing that runs after accelerated growth is calculated
}

