% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/util.R
\name{df_sorter}
\alias{df_sorter}
\title{df_sorter}
\usage{
df_sorter(x, by = 1, decreasing = FALSE, ...)
}
\arguments{
\item{x}{data frame to sort}

\item{by}{how to sort}

\item{decreasing}{logical toggle to change sort order}

\item{...}{additional arguments}
}
\description{
helper function used by report_dispatcher
}

