% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/norm_plots.R
\name{empty_norm_grade_space}
\alias{empty_norm_grade_space}
\title{empty norm grade space}
\usage{
empty_norm_grade_space(measurementscale, trace_lines = c(5, 10, 20, 30, 40,
  50, 60, 70, 80, 90, 95))
}
\arguments{
\item{measurementscale}{a NWEA map measurementscale}

\item{trace_lines}{vector of percentiles to show.  must be between 1 and 99.}
}
\description{
shows the norm space across grade levels or a given subject
}

