% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ensure_fields}
\alias{ensure_fields}
\title{ensure_fields}
\usage{
ensure_fields(fields_vector, df)
}
\arguments{
\item{fields_vector}{vector of fields that your function needs.}

\item{df}{data frame that needs to have the fields}
}
\description{
a simple wrapper around ensurer to check for necessary fields
used by a function
}
