% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/goals_and_projections.R
\name{ensure_goals_obj}
\alias{ensure_goals_obj}
\title{ensure_goals_obj}
\usage{
ensure_goals_obj(.)
}
\arguments{
\item{.}{dot-placeholder, per ensurer doc.}
}
\description{
a contract that ensures that a goal object's has the proper
elements and that the \code{goals} element data frame has columns names
\code{accel_growth} and \code{met_acc}
}

