% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapvizieR_object.R
\name{ensure_is_mapvizieR}
\alias{ensure_is_mapvizieR}
\title{ensure_is_mapvizieR}
\usage{
ensure_is_mapvizieR(x)
}
\arguments{
\item{x}{object to check}
}
\description{
a contract that ensures that an object is a mapvizieR object at runtime.
}
