% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roster_check.R
\name{ensure_roster_names}
\alias{ensure_roster_names}
\title{ensure_roster_names}
\usage{
ensure_roster_names(.)
}
\arguments{
\item{.}{dot-placeholder, per ensurer doc.}
}
\description{
does the roster have a studentid and grade field?
}
