% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roster_check.R
\name{ensure_roster_types}
\alias{ensure_roster_types}
\title{ensure_roster_types}
\usage{
ensure_roster_types(roster)
}
\arguments{
\item{roster}{roster data frame to check}
}
\description{
checks if certain roster fields are of the correct type
}
