% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ensure_rows_in_df}
\alias{ensure_rows_in_df}
\title{ensure_rows_in_df}
\usage{
ensure_rows_in_df(df)
}
\arguments{
\item{df}{data frame to check}
}
\description{
a contract that verifies that a data set isn't length zero
}
