% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzz_test.R
\name{error_ridden_plot}
\alias{error_ridden_plot}
\title{error ridden plot}
\usage{
error_ridden_plot(mapvizieR_obj, studentids)
}
\arguments{
\item{mapvizieR_obj}{a conforming mapvizieR object}

\item{studentids}{a vector of studentids}
}
\description{
a plot that is definitely going to break
}

