% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_RIT.R
\name{estimate_rit}
\alias{estimate_rit}
\title{estimate RIT score for a student using regression or interpolation}
\usage{
estimate_rit(mapvizieR_obj, studentid, measurementscale, target_date,
  method = c("closest", "lm", "interpolate"), num_days = 180,
  forward = TRUE)
}
\arguments{
\item{mapvizieR_obj}{mapvizieR object}

\item{studentid}{target studentid}

\item{measurementscale}{target subject}

\item{target_date}{date of interest, \code{Y-m-d} format}

\item{method}{which method to use to estimate RIT score}

\item{num_days}{function will only return test score within num_days of target_date}

\item{forward}{default is TRUE, set to FALSE if only scores before target_date should be chosen for 'closest' method}
}
\description{
Given studentid, measurementscale, and a target_date, the 
function will return an estimated score based on selected method
}

