% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cgp_prep.R
\name{eval_cgp_result}
\alias{eval_cgp_result}
\title{eval_cgp_result}
\usage{
eval_cgp_result(measurementscale, grade, growth_window, baseline_avg_rit = NA,
  baseline_avg_npr = NA, tolerance = 10,
  sch_growth_study = sch_growth_norms_2012, calc_for = c(1:99))
}
\description{
given a growth window and observed rit/npr change, what cohort
backwards looking oppo
}

