% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_data.R
\docType{data}
\name{ex_CombinedStudentsBySchool}
\alias{ex_CombinedStudentsBySchool}
\title{example ex_CombinedStudentsBySchool file (formerly Comprehensive Data File)}
\format{\describe{
\item{TermName}{TermName}
\item{DistrictName}{DistrictName}
\item{SchoolName}{SchoolName}
\item{StudentLastName}{StudentLastName}
\item{StudentFirstName}{StudentFirstName}
\item{StudentMI}{StudentMI}
\item{StudentID}{StudentID}
\item{StudentDateOfBirth}{StudentDateOfBirth}
\item{StudentEthnicGroup}{StudentEthnicGroup}
\item{StudentGender}{StudentGender}
\item{Grade}{Grade}
}}
\source{
NWEA MAP partner support team (thanks!)
}
\usage{
ex_CombinedStudentsBySchool
}
\description{
Roster and demographic records, in the format provided by web-based MAP.
Data provided from NWEA, by request.  
(Note that this is training data, and these students/student records *are not real*!)
}
\keyword{datasets}

