% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/util.R
\name{extract_academic_year}
\alias{extract_academic_year}
\title{extract_academic_year}
\usage{
extract_academic_year(x)
}
\arguments{
\item{x}{a data frame}
}
\value{
data frame with disambiguated termname and map_year_academic
}
\description{
\code{extract_academic_year} break up the termname field into year and fallwinterspring
}

