% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fall_goals_report.R
\name{fall_goals_data_table}
\alias{fall_goals_data_table}
\title{Basline and Goals, for printing reports}
\usage{
fall_goals_data_table(mapvizieR_obj, studentids, measurementscale, start_fws,
  start_year_offset, end_fws, end_academic_year, end_grade,
  start_fws_prefer = NA, calc_for = 80, output = "both", font_size = 20)
}
\arguments{
\item{mapvizieR_obj}{a valid mapvizieR object}

\item{studentids}{vector of studentids}

\item{measurementscale}{target subject}

\item{start_fws}{character, starting season for school growth norms}

\item{start_year_offset}{0 if start season is same, -1 if start is prior year.}

\item{end_fws}{ending season}

\item{end_academic_year}{ending academic year}

\item{end_grade}{grade level at end of growth window (for lookup)}

\item{start_fws_prefer}{if more than one start_fws, what is the preferred term?}

\item{calc_for}{what CGPs to calc for?  vector of integers between 1:99}

\item{output}{c('both', 'baseline', 'goals')}

\item{font_size}{how big to print}
}
\value{
gridArrange object of grobs
}
\description{
Basline and Goals, for printing reports
}
