% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fall_goals_report.R
\name{fall_goals_report}
\alias{fall_goals_report}
\title{Title}
\usage{
fall_goals_report(mapvizieR_object, studentids, context, start_fws = "Spring",
  start_year_offset = -1, end_fws = "Spring", end_academic_year = 2015)
}
\arguments{
\item{mapvizieR_object}{a valid mapvizieR_object}

\item{studentids}{a vector of studentids to run}

\item{context}{what school/grade/class/etc grouping is represented?}

\item{start_fws}{character, starting season for school growth norms}

\item{start_year_offset}{0 if start season is same, -1 if start is prior year.}

\item{end_fws}{ending season}

\item{end_academic_year}{ending academic year}
}
\value{
a multipage report, represented as a list of grobs.
}
\description{
Title
}

