% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/util.R
\name{fall_spring_me}
\alias{fall_spring_me}
\title{Fall-Spring Me}
\usage{
fall_spring_me(grade_season)
}
\arguments{
\item{grade_season}{a grade level between -1 and 12}
}
\value{
a labeled string
}
\description{
\code{fall_spring_me} tranforms grade levels into labels for charts; eg 4.2 -> F5
}

