% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_wrappers.R
\name{filter_.mapvizieR_data}
\alias{filter_.mapvizieR_data}
\title{filter wrapper}
\usage{
filter_.mapvizieR_data(df, ...)
}
\arguments{
\item{df}{data.frame}

\item{...}{additional args}
}
\value{
data.frame
}
\description{
wrapper for filter that preserves classes of data frames
}
