% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/find_duplicate_ids.R
\name{find_duplicate_ids}
\alias{find_duplicate_ids}
\title{Find's duplicate roster ids and attempts to fix them by collapsing}
\usage{
find_duplicate_ids(students_by_school)
}
\arguments{
\item{students_by_school}{the StudentsBySchool roster as a data.frame}
}
\value{
a data.frame if collapsible, otherwise a message
}
\description{
Find's duplicate roster ids and attempts to fix them by collapsing
}

