% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{force_string_breaks}
\alias{force_string_breaks}
\title{force_string_breaks}
\usage{
force_string_breaks(string, n_char)
}
\arguments{
\item{string}{the string}

\item{n_char}{break every x}
}
\description{
a text processing function to insert a \\n break every x characters
}

