% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzz_test.R
\name{fuzz_test_plot}
\alias{fuzz_test_plot}
\title{fuzz_test_plot}
\usage{
fuzz_test_plot(
  plot_name,
  n = 100,
  additional_args = list(),
  mapvizieR_obj = mapvizieR(cdf = ex_CombinedAssessmentResults, roster =
    ex_CombinedStudentsBySchool)
)
}
\arguments{
\item{plot_name}{name of a plot, as text.  gets thrown to do call}

\item{n}{how many times to test?}

\item{additional_args}{all plots will get "mapvizieR_obj" and "studentids".  if your
plot needs additional args, pass them here.}

\item{mapvizieR_obj}{a \code{\link{mapvizieR}} object.}
}
\description{
throws a random subset of students against a plot n times, and reports back
if a valid ggplot gets returned
}
