% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cgp_prep.R
\name{gain_needed}
\alias{gain_needed}
\title{gain_needed}
\usage{
gain_needed(percentile, sd_gain, mean_gain)
}
\arguments{
\item{percentile}{growth percentile, between 0-100}

\item{sd_gain}{sd for population growth}

\item{mean_gain}{typical growth for population}
}
\description{
rit gain needed to reach given percentile
}

