% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/galloping_elephants.R
\name{galloping_elephants}
\alias{galloping_elephants}
\title{RIT distribution change (affectionately titled 'Galloping Elephants')}
\usage{
galloping_elephants(data, first_and_spring_only = TRUE,
  entry_grades = c(-0.8, 4.2))
}
\arguments{
\item{first_and_spring_only}{show all terms, or only entry & spring?  default is TRUE.}

\item{cdf_long_prepped}{cdf.  run \code{prep_cdf_long} to get into proper format.}

\item{entry_grades.}{which grades are entry grades?}
}
\value{
a ggplot object.
}
\description{
\code{galloping_elephants} returns ggplot density distributions that show change
 in RIT over time
}

