% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/growth_df_prep.R
\name{generate_growth_df}
\alias{generate_growth_df}
\title{generate_growth_df}
\usage{
generate_growth_df(prepped_cdf, start_season, end_season,
  norm_df = norms_students_2011, include_unsanctioned_windows = FALSE)
}
\arguments{
\item{prepped_cdf}{a conforming prepped_cdf data frame}

\item{start_season}{the start of the growth window ("Fall", "Winter", or "Spring")}

\item{end_season}{the end of the growth window ("Fall", "Winter", or "Spring")}

\item{norm_df}{defaults to norms_students_2011.  if you have a conforming norms object,
you can use generate_growth_df to produce a growth data frame for those norms.
example usage: calculate college ready growth norms, and use generate_growth_df to see
if students met them.}

\item{include_unsanctioned_windows}{if TRUE, generate_growth_df will
return some additional growth windows like 'Spring to Winter', which aren't in the
official norms (but might be useful for progress monitoring).}
}
\value{
a data frame with all rows where the student had at least ONE matching
test event (start or end)
}
\description{
\code{generate_growth_df} takes a CDF and given two seasons (start and end) saturates
all possible growth calculations for a student and returns a long data frame with the
results.
}
\details{
This function returns a growth data frame, with one row per student per test per
valid 'growth_window', such as 'Fall to Spring'.
}

