% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/growth_df_prep.R
\name{generate_growth_dfs}
\alias{generate_growth_dfs}
\title{generate_growth_df}
\usage{
generate_growth_dfs(processed_cdf,
  norm_df_long = norms_students_wide_to_long(norms_students_2011),
  include_unsanctioned_windows = FALSE)
}
\arguments{
\item{processed_cdf}{a conforming processed_cdf data frame}

\item{norm_df_long}{defaults to norms_students_2011.  if you have a conforming norms object,
you can use generate_growth_df to produce a growth data frame for those norms.
example usage: calculate college ready growth norms, and use generate_growth_df to see
if students met them.}

\item{include_unsanctioned_windows}{if TRUE, generate_growth_df will
return some additional growth windows like 'Spring to Winter', which aren't in the
official norms (but might be useful for progress monitoring).}
}
\value{
a data frame with all rows where the student had at least ONE matching
test event (start or end)
}
\description{
\code{generate_growth_df} takes a CDF and given two seasons (start and end) saturates
all possible growth calculations for a student and returns a long data frame with the
results.
}
\details{
A workflow wrapper that calls a variety of growth_df prep functions.
Given a mapvizieR processed cdf, this function will return a a growth
data frame, with one row per student per test per valid 'growth_window',
eg 'Fall to Spring'.
}

