% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goals_and_projections.R
\name{goal_kipp_tiered}
\alias{goal_kipp_tiered}
\title{Create KIPP Tiered accelerated growth goals}
\usage{
goal_kipp_tiered(mapvizier_object, iterations = 1)
}
\arguments{
\item{mapvizier_object}{a \code{\link{mapvizieR}} object.}

\item{iterations}{the number of iterations out from any student test event
you wish to continue projecting student growth.  This features is not 
yet implemented, so it only projects growth one iteration.}
}
\description{
\code{goal_kipp_tiered} is a "goal function": it creates a list with three
elements: a \code{goals} data frame (including  fields \code{accel_growth} and 
\code{met_accel_growth} used in the \code{growth_df} of a \code{\link{mapvizieR}}
object.
}
\examples{
\dontrun{
data(ex_CombinedAssessmentResults)
data(ex_CombinedStudentsBySchool)

cdf_mv <- mapvizieR(ex_CombinedAssessmentResults, 
                    ex_CombinedStudentsBySchool)
                    
goals<-goal_kipp_tiered(cdf_mv)                     
}
}
