% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strand_summary_plot.R
\name{goal_strand_summary_plot}
\alias{goal_strand_summary_plot}
\title{Plot MAP Goal Strand Summary Info}
\usage{
goal_strand_summary_plot(
  mapvizieR_obj,
  studentids,
  measurementscale,
  fws = c("Fall", "Winter", "Spring"),
  year,
  cohort = FALSE,
  spring_is_first = FALSE,
  filter_args = NULL
)
}
\arguments{
\item{mapvizieR_obj}{a \code{mapvizieR} object}

\item{studentids}{vector of student id numbers for students to plot}

\item{measurementscale}{measurementscale to plot}

\item{fws}{seasons (fall, winter, or spring) as a character vector to plot}

\item{year}{academic year to plot}

\item{cohort}{cohort year to plot as integer or FALSE (the default).  If `cohort`
is not FALSE then `year` is ignored}

\item{spring_is_first}{logical indicating whether spring is in the given 
academic year or from the year prior.}

\item{filter_args}{a list of character vectors to filter `mapvizieR_obj$cdf` 
by that is passed to \code{\link[dplyr]{filter}}}
}
\value{
a ggplot2 object
}
\description{
Plots a group of students' average RIT scores and RIT Ranges for 
goal strands.
}
\details{
Creates and prints a ggplot2 object showing average and and average range 
of goal strand RIT scores
}
\examples{
\dontrun{
require(dplyr)

data("ex_CombinedStudentsBySchool")
data("ex_CombinedAssessmentResults")

map_mv <- mapvizieR(ex_CombinedAssessmentResults, ex_CombinedStudentsBySchool)

ids <- ex_CombinedStudentsBySchool \%>\%
dplyr::filter(
  TermName == "Spring 2013-2014") \%>\% select(StudentID) \%>\%
  unique()

goal_strand_summary_plot(map_mv, 
   ids$StudentID, 
   measurementscale = "Reading", 
   year = 2013, 
   cohort = 2019, 
   fws  = c("Winter", "Spring")
   )
}
}
