% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/goalbar.R
\name{goalbar}
\alias{goalbar}
\title{goal_bar}
\usage{
goalbar(mapvizieR_obj, studentids, measurementscale, start_fws,
  start_academic_year, end_fws, end_academic_year,
  goal_labels = c(accel_growth = "Made Accel Growth", typ_growth =
  "Made Typ Growth", positive_but_not_typ = "Below Typ Growth", negative_growth
  = "Negative Growth", no_start = sprintf("Untested: \%s", start_fws), no_end =
  sprintf("Untested: \%s", end_fws)), goal_colors = c("#CC00FFFF",
  "#0066FFFF", "#CCFF00FF", "#FF0000FF", "#FFFFFF", "#F0FFFF"),
  inprogress_prorater = NA, complete_obsv = FALSE)
}
\arguments{
\item{mapvizieR_obj}{mapvizieR object}

\item{studentids}{target students}

\item{measurementscale}{target subject}

\item{start_fws}{starting season}

\item{start_academic_year}{starting academic year}

\item{end_fws}{ending season}

\item{end_academic_year}{ending academic year}

\item{goal_labels}{what labels to show for each goal category.  must be in order from
highest to lowest.}

\item{goal_colors}{what colors to show for each goal category}

\item{inprogress_prorater}{default is NA.  if set to a decimal value, what percent of the goal
is considered ontrack?}

\item{complete_obsv}{if TRUE, limit only to students who have BOTH a start
and end score. default is FALSE.}

\item{subgroup_cols}{what subgroups in mapvizier roster do you want to cut by?  default
is starting_quartile}

\item{pretty_names}{nicely formatted names for the column cuts used above.}
}
\description{
a simple bar chart that shows the percentage of a cohort at different goal
states (met / didn't meet)
}

