% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/util.R
\name{grade_level_season}
\alias{grade_level_season}
\title{grade_level_season}
\usage{
grade_level_season(season)
}
\arguments{
\item{season}{'Fall', 'Winter', 'Spring', or 'Summer'.}
}
\value{
decimal offset that matches the season
}
\description{
returns appropriate offset given a season
}

