% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf_prep.R
\name{grade_level_seasonify}
\alias{grade_level_seasonify}
\title{grade_level_seasonify}
\usage{
grade_level_seasonify(cdf)
}
\arguments{
\item{cdf}{a cdf that has 'grade' and 'fallwinterspring' columns (eg product of )
\code{grade_levelify()}}
}
\value{
a data frame with a 'grade_level_season' column
}
\description{
\code{grade_level_seasonify} turns grade level into a simplified continuous scale, 
using consistent offsets for MAP 'seasons'
}

