% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapvizieR_object.R
\name{grade_levelify_cdf}
\alias{grade_levelify_cdf}
\title{grade_levelify_cdf}
\usage{
grade_levelify_cdf(prepped_cdf, roster)
}
\arguments{
\item{prepped_cdf}{a cdf file that passes the checks in \code{check_cdf_long}}

\item{roster}{a roster that passes the checks in \code{check_roster}}
}
\value{
a vector of grades
}
\description{
\code{grade_levelify_cdf} adds a student's grade level at test time to
the cdf.  grade level is required for a variety of growth calculations.
}
