% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf_prep.R
\name{grade_season_factors}
\alias{grade_season_factors}
\title{grade_season_factors}
\usage{
grade_season_factors(x)
}
\arguments{
\item{x}{a cdf that has grade_level_season and grade_season_label}
}
\description{
helper function that 1) converts grade_season_label to factor and
2) orders the labels based on grade_level_season
}

