% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cdf_prep.R
\name{grade_season_labelify}
\alias{grade_season_labelify}
\title{grade_season_labelify}
\usage{
grade_season_labelify(x)
}
\arguments{
\item{x}{a cdf that has 'grade_level_season' (eg product of grade_level_seasonify)
\code{grade_levelify()}}
}
\value{
a data frame with a grade_season_labels
}
\description{
\code{grade_season_labelify} returns an abbreviated label ('5S') that is useful when
labelling charts
}

