% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mapvizieR_object.R
\name{grade_season_sortify}
\alias{grade_season_sortify}
\title{grade_season_sortify}
\usage{
grade_season_sortify(x)
}
\arguments{
\item{x}{a cdf that has 'grade_level_season' (eg product of grade_level_seasonify)
\code{grade_levelify()}}
}
\value{
a data frame with a grade_season_sorted
}
\description{
\code{grade_season_sortify} returns a sortable grade season label
}

