% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_util.R
\name{grob_justifier}
\alias{grob_justifier}
\title{grob_justifier}
\usage{
grob_justifier(x, hjust = "center", vjust = "center")
}
\arguments{
\item{x}{a grob}

\item{hjust}{left, center, or right}

\item{vjust}{top, center, or bottom}
}
\description{
a great little function that baptiste wrote up on my SO queston: 
http://stackoverflow.com/a/25456672/561698
}
