% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_wrappers.R
\name{group_by_.mapvizieR_data}
\alias{group_by_.mapvizieR_data}
\title{group_by wrapper}
\usage{
group_by_.mapvizieR_data(df, ...)
}
\arguments{
\item{df}{data.frame}

\item{...}{additional args}
}
\value{
data.frame
}
\description{
wrapper for group_by that preserves classes of data frames
}

