% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/growth_df_prep.R
\name{growth_norm_lookup}
\alias{growth_norm_lookup}
\title{growth_norm_lookup}
\usage{
growth_norm_lookup(incomplete_growth_df, processed_cdf, norm_df_long,
  include_unsanctioned_windows, ...)
}
\arguments{
\item{incomplete_growth_df}{a growth df in process.  needs to have growth
windows, start_grade, and start_testritscore.}

\item{processed_cdf}{conforming mapvizieR processed cdf.  needed for
unsanctioned windows.}

\item{norm_df_long}{a data frame of normative expectations}

\item{include_unsanctioned_windows}{check generate_growth_df for a
description.}

\item{...}{currently not used.}
}
\description{
called by \code{generate_growth_df} to return
growth norms for growth data frames in process
}

