% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/growth_status.R
\name{growth_status_scatter}
\alias{growth_status_scatter}
\title{growth status scatter}
\usage{
growth_status_scatter(mapvizieR_obj, studentids, measurementscale_in, fws,
  academic_year)
}
\arguments{
\item{mapvizieR_obj}{mapvizieR object}

\item{studentids}{target students}

\item{measurementscale_in}{target subject}

\item{fws}{season}

\item{academic_year}{academic year}
}
\value{
a ggplot object
}
\description{
growth status scatter
}

