% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/growth_df_prep.R
\name{growth_testid_lookup}
\alias{growth_testid_lookup}
\title{growth_testid_lookup}
\usage{
growth_testid_lookup(scaffold, processed_cdf)
}
\arguments{
\item{scaffold}{output of \code{build_growth_scaffolds}}

\item{processed_cdf}{conforming mapvizieR processed cdf}
}
\description{
a helper function for \code{generate_growth_df}
given a scaffold of student/season growth windows, finds the
matching test data in the cdf
}

