% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_util.R
\name{h_var}
\alias{h_var}
\title{h_var}
\usage{
h_var(text, size, gp = grid::gpar(fontsize = size, fontface = "bold"))
}
\arguments{
\item{text}{text you want on the grob}

\item{size}{font size}

\item{gp}{graphics paramaters.}
}
\description{
utility function to make a text grob of desired size
}

