% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/haid_plot.R
\name{haid_plot}
\alias{haid_plot}
\title{Chris Haid's Waterfall-Rainbow-Arrow Chart}
\usage{
haid_plot(mapvizieR_obj, studentids, measurementscale, start_fws,
  start_academic_year, end_fws, end_academic_year,
  sort_column = "start_testritscore", p_growth_colors = c("tan4", "snow4",
  "gold", "red"), p_growth_tiers = c("Positive", "Typical", "College Ready",
  "Negative"), p_quartile_colors = c("#f3716b", "#79ac41", "#1ebdc2",
  "#a57eb8"), p_name_size = 3, p_alpha = 1, p_name_offset = 0.04)
}
\arguments{
\item{mapvizieR_obj}{mapvizieR object}

\item{studentids}{target students}

\item{measurementscale}{target subject}

\item{start_fws}{starting season}

\item{start_academic_year}{starting academic year}

\item{end_fws}{ending season}

\item{end_academic_year}{ending academic year}

\item{p_growth_colors}{vector of colors passed to ggplot.}

\item{p_growth_tiers}{vector of colors passed to ggplot.}

\item{p_quartile_colors}{vector of colors passed to ggplot.}

\item{p_name_size}{sets point size of student names.}

\item{p_alpha}{sets level fo transparency for goals.}

\item{p_name_offset}{percentage to offset for negative names.  make bigger if plot
is smaller.}
}
\value{
prints a ggplot object
}
\description{
\code{haid_plot} returns a ggplot object showing student MAP performance
for a group of students
}
\details{
This function builds and prints a graphic that plots MAP performance
over one or two seasons. RIT scores are color coded by percentile.
}

