% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/historic_nth_percentile_plot.R
\name{historic_nth_percentile_plot}
\alias{historic_nth_percentile_plot}
\title{Percent of students above a given percentile, by cohort}
\usage{
historic_nth_percentile_plot(mapvizieR_obj, studentids, measurementscale,
  target_percentile = 75, first_and_spring_only = TRUE,
  entry_grade_seasons = c(-0.8, 4.2), small_n_cutoff = 0.5)
}
\arguments{
\item{mapvizieR_obj}{a conforming mapvizieR object}

\item{studentids}{vector of studentids}

\item{measurementscale}{target subject}

\item{target_percentile}{what is the goal percentile for calcs (pct
of students at/above this percentile?)}

\item{first_and_spring_only}{logical, should we drop winter/fall scores?}

\item{entry_grade_seasons}{what seasons are entry grades?}

\item{small_n_cutoff}{any cohort below this percent will get filtered out.
default is 0.5, eg cohorts under 0.5 of max size will get dropped.}
}
\value{
a ggplot object
}
\description{
Percent of students above a given percentile, by cohort
}

