% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historic_recap_report.R
\name{historic_recap_report}
\alias{historic_recap_report}
\title{Historic Recap Report}
\usage{
historic_recap_report(mapvizieR_obj, studentids, measurementscale,
  target_percentile = 75, first_and_spring_only = TRUE,
  entry_grade_seasons = c(-0.8, 4.2), small_n_cutoff = 0.2,
  min_cohort_size = -1, title_text = "")
}
\arguments{
\item{mapvizieR_obj}{conforming mapvizieR object}

\item{studentids}{vector of studentids}

\item{measurementscale}{target subject}

\item{target_percentile}{integer, what is the 'goal' percentile to show
progress against?}

\item{first_and_spring_only}{logical, should we drop fall and winter 
scores?}

\item{entry_grade_seasons}{numeric vector, what are 'entry' grades to the school?}

\item{small_n_cutoff}{numeric, drop observations that are smaller than X% of the
cohort maximum.}

\item{min_cohort_size}{integer, filter out cohorts with less than this many students
in them.}

\item{title_text}{report title}
}
\value{
grob, with plots arranged
}
\description{
looks back across multiple cohorts.  
summarizes growth and and attainment
}

