% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf_prep.R
\name{id_cdf_type}
\alias{id_cdf_type}
\title{identify type of cdf from AssessmentResults file.}
\usage{
id_cdf_type(cdf)
}
\arguments{
\item{cdf}{an Assessment Results data frame.}
}
\value{
one of "Client-Server", "WBM pre-2015", "WBM post-2015"
}
\description{
identify type of cdf from AssessmentResults file.
}
\examples{
data(ex_CombinedAssessmentResults)
id_cdf_type(ex_CombinedAssessmentResults)

data(ex_CombinedAssessmentResults_pre_2015)
id_cdf_type(ex_CombinedAssessmentResults_pre_2015)
}

