% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_rit.R
\name{imputation_grouper}
\alias{imputation_grouper}
\title{Utility function to identify groups/runs when imputing}
\usage{
imputation_grouper(logicals)
}
\arguments{
\item{logicals}{a vector of logicals (indicating if the rit score is known or NA)}
}
\value{
a vector of integers, representing the sequential group number
}
\description{
Utility function to identify groups/runs when imputing
}

