% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_rit.R
\name{imputation_scaffold}
\alias{imputation_scaffold}
\title{Build out the base scaffold of possible terms for every student.}
\usage{
imputation_scaffold(cdf, interpolate_only = TRUE)
}
\arguments{
\item{cdf}{a processed cdf.  assumes that there are no same student/subj/season 
dupes.}

\item{interpolate_only}{should the scaffold return ALL seasons, ever, or only
ones in between the student's first/last test?}
}
\value{
a cdf, with rows for imputation
}
\description{
Build out the base scaffold of possible terms for every student.
}
