% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_rit.R
\name{impute_rit}
\alias{impute_rit}
\title{Impute missing RIT scores}
\usage{
impute_rit(mapvizieR_obj, studentids, measurementscale,
  impute_method = "simple_average", interpolate_only = TRUE)
}
\arguments{
\item{mapvizieR_obj}{a mapvizieR object}

\item{studentids}{a vector of studentids to run}

\item{measurementscale}{desired subject}

\item{impute_method}{one of: c('simple_average')}

\item{interpolate_only}{should the scaffold return ALL seasons, ever, or only
ones in between the student's first/last test?}
}
\value{
a cdf object, with imputed rows
}
\description{
Impute missing RIT scores
}
