% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapvizieR_object.R
\name{is.mapvizieR}
\alias{is.mapvizieR}
\title{Reports whether x is a mapvizier object}
\usage{
is.mapvizieR(x)
}
\arguments{
\item{x}{an object to test}
}
\description{
Reports whether x is a mapvizier object
}
