% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{kipp_quartile}
\alias{kipp_quartile}
\title{Calcualte KIPP Foundation style quartiles from percentile vector}
\usage{
kipp_quartile(x, return_factor = TRUE, proper_quartile = FALSE)
}
\arguments{
\item{x}{vector of percentiles to be converted to quartiels}

\item{return_factor}{default is \code{TRUE}.  If set to \code{FALSE} returns integers rather than factors.}

\item{proper_quartile}{defaul is \code{FALSE}.  If set to \code{TRUE} returns traditional quartiles rather then KIPP Foundation quartiles.}
}
\value{
a vector of \code{length(x)}.
}
\description{
\code{kipp_quartile} returns an integer or factor vector quartiles.
}
\details{
This function calculates the KIPP Foundation's (kinda fucked up) quartile (i.e., the foundation
breaks with stanard mathematical pracitce and puts the 50th percenile
in the  3rd rather than the 2nd quartile). It takes a vector of percentiles and translates 
those into quartiles, where the 25th, 50th, and 75th percentils are shifted up 
into the 2nd, 3rd, and 4th quartiles, respectively. You can revert to a traditional 
quartile calculation by setting the \code{proper.quartile} argument to \code{TRUE}.
}
\examples{
x <- sample(x=1:99, 100,replace = TRUE)
kipp_quartile(x)
kipp_quartile(x, proper_quartile=TRUE)
kipp_quartile(x, proper_quartile=TRUE, return_factor=FALSE)
}

