% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/typ_growth_distro.R
\name{kipp_typ_growth_distro}
\alias{kipp_typ_growth_distro}
\title{KIPP % Typ Growth Network Distribution}
\usage{
kipp_typ_growth_distro(nat_results_df, measurementscale, academic_year,
  grade_level, start_fws, end_fws, comparison_name, comparison_pct_typ_growth,
  replace_nat_results_match = FALSE, de_kippify_names = TRUE,
  de_schoolify_names = TRUE)
}
\arguments{
\item{nat_results_df}{KIPP results nationwide.  Ask R&E for the raw data file - same as the data
in the HSR tableau reports.}

\item{measurementscale}{MAP subject}

\item{academic_year}{the academic year}

\item{grade_level}{comparison grade level}

\item{start_fws}{starting season}

\item{end_fws}{ending season}

\item{comparison_name}{this school name}

\item{comparison_pct_typ_growth}{pct keep up, this school}

\item{replace_nat_results_match}{if using last year's data, remove this school's name}

\item{de_kippify_names}{shorten names by removing KIPP prefix}

\item{de_schoolify_names}{shorten names by removing 'Academy', 'Primary', etc.}

\item{start_academic_year}{starting academic year}

\item{end_academic_year}{ending academic year}
}
\value{
a ggplot chart
}
\description{
shows your school relative to KIPP schools nationwide (if you have that dataset :))
}

