% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_pager.R
\name{knj_two_pager}
\alias{knj_two_pager}
\title{two-pager, KNJ style}
\usage{
knj_two_pager(mapvizieR_obj, studentids, measurementscale, end_fws,
  end_academic_year, detail_academic_year, candidate_start_fws = c("Fall",
  "Spring"), start_year_offsets = c(0, -1), prefer_fws = "Spring",
  national_data_frame = NA, title_text = "", ...)
}
\arguments{
\item{mapvizieR_obj}{mapvizieR object}

\item{studentids}{target students}

\item{measurementscale}{target subject}

\item{end_fws}{ending season}

\item{end_academic_year}{ending academic year}

\item{detail_academic_year}{don't mask any data for this academic year}

\item{candidate_start_fws}{what possible start terms should we consider?}

\item{start_year_offsets}{using same order as candidate_start_fws, if we pick this year, what should we add
to the end_academic_year to get the right year's data?  eg, -1 for spring/spring.}

\item{prefer_fws}{what is the 'best' start term?}

\item{national_data_frame}{for internal KIPP use - a data frame showing % making
typ growth across KIPP}

\item{title_text}{what is this report called?}

\item{...}{additional arguments}
}
\value{
a ggplot object
}
\description{
similar to the existing two pager, but attempts to auto-set fall-spring vs 
spring/spring
}

