% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/util.R
\name{lower_df_names}
\alias{lower_df_names}
\title{lower_df_names}
\usage{
lower_df_names(x)
}
\arguments{
\item{x}{a data frame}
}
\value{
data frame with modified names
}
\description{
\code{lower_df_names} a utility function to take a data frame and return it with
lowercase names
}

