% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cgp_prep.R
\name{mapviz_cgp}
\alias{mapviz_cgp}
\title{mapvizieR interface to simplify CGP calculations}
\usage{
mapviz_cgp(mapvizieR_obj, studentids, measurementscale, start_fws,
  start_academic_year, end_fws, end_academic_year)
}
\arguments{
\item{mapvizieR_obj}{mapvizieR object}

\item{studentids}{target students}

\item{measurementscale}{target subject}

\item{start_fws}{starting season}

\item{start_academic_year}{starting academic year}

\item{end_fws}{ending season}

\item{end_academic_year}{ending academic year}
}
\description{
mapvizieR interface to simplify CGP calculations
}

