% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgp_prep.R
\name{mapviz_cgp_targets}
\alias{mapviz_cgp_targets}
\title{mapvizieR interface to simplify growth goal calculations}
\usage{
mapviz_cgp_targets(mapvizieR_obj, studentids, measurementscale, start_fws,
  start_year_offset, end_fws, end_academic_year, end_grade,
  start_fws_prefer = NA)
}
\arguments{
\item{mapvizieR_obj}{mapvizieR object}

\item{studentids}{target students}

\item{measurementscale}{target subject}

\item{start_fws}{one academic season (if known); 
pass vector of two and mapviz_cgp_targets will pick}

\item{start_year_offset}{0 if start season is same, -1 if start is prior year.}

\item{end_fws}{ending season}

\item{end_academic_year}{ending academic year}

\item{start_fws_prefer}{which term is preferred? not required if only one start_fws is passed}

\item{specify}{the ending grade for the growth (this can't be reliably
inferred from data).}
}
\value{
data frame of cgp targets
}
\description{
given an explicit window, or a composite baseline, will calculate
CGP targets
}

