% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapvizieR-package.R
\docType{package}
\name{mapvizieR-package}
\alias{mapvizieR-package}
\title{mapvizieR: Visualization and Data Analysis Tools for NWEA MAP Student Data}
\description{
A suite of analytic, visualization, and reporting tools for NWEA MAP Data. Provides functions for creating visualizations of student growth, cohort status, and goal attainment from MAP assessment data. Includes specialized plots like the becca plot, galloping elephants, haid plot, and quealy subgroups visualization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/almartin82/mapvizieR}
  \item \url{https://almartin82.github.io/mapvizieR}
  \item Report bugs at \url{https://github.com/almartin82/mapvizieR/issues}
}

}
\author{
\strong{Maintainer}: Andrew Martin \email{almartin@gmail.com}

Authors:
\itemize{
  \item Chris Haid \email{chaid@kippchicago.org}
}

}
\keyword{internal}
