% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mapvizier.R
\name{mapvizier_colors}
\alias{mapvizier_colors}
\alias{mapvizier_quartile_colors}
\alias{mapvizier_growth_colors}
\alias{mapvizier_kipp_colors}
\title{mapvizieR Color Palettes}
\usage{
mapvizier_quartile_colors()

mapvizier_growth_colors()

mapvizier_kipp_colors()
}
\value{
A character vector of color hex codes
}
\description{
Standard color palettes used throughout mapvizieR visualizations.
}
\examples{
mapvizier_quartile_colors()
mapvizier_growth_colors()

}
