% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_wrappers.R
\name{mapvizieR_data}
\alias{mapvizieR_data}
\title{makes data a mapvizieR_data object}
\usage{
mapvizieR_data(df)
}
\arguments{
\item{df}{a data frame}
}
\description{
I have absolutely no idea what I am doing.gif
}

