% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf_prep.R
\name{migrate_cdf_to_2015_std}
\alias{migrate_cdf_to_2015_std}
\title{Migrate pre-2015 CDFs  (both client-server nad WBM) to post-2015 specification,}
\usage{
migrate_cdf_to_2015_std(cdf)
}
\arguments{
\item{cdf}{the Assessment Results table from a Comprehensive Data file as a data.frame}
}
\value{
a data.frame of the Assessment Results table in the post-2015 CDF format
}
\description{
Migrate pre-2015 CDFs  (both client-server nad WBM) to post-2015 specification,
}

