% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{min_subgroup_filter}
\alias{min_subgroup_filter}
\title{min_subgroup_filter}
\usage{
min_subgroup_filter(df, subgroup_name, small_n_cutoff = -1)
}
\arguments{
\item{df}{some data frame}

\item{small_n_cutoff}{anything below this percent will get filtered out.
default is -1, eg off}

\item{subgroup}{name of a column of the data frame}
}
\description{
given a data frame and some arbitrary subgroup,
return only the rows that are members of subgroups that make up
at least n % of the total data frame
}

