% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{min_term_filter}
\alias{min_term_filter}
\title{min_term_filter}
\usage{
min_term_filter(cdf, small_n_cutoff = -1)
}
\arguments{
\item{cdf}{conforming cdf}

\item{small_n_cutoff}{anything below this percent will get filtered out.
default is -1, eg off}
}
\description{
returns only grade_season data where the number of students represented is at
least N% of the total
}

