% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_cgp_hist.R
\name{multi_cohort_cgp_hist_plot}
\alias{multi_cohort_cgp_hist_plot}
\title{Multiple Cohort CGP histories}
\usage{
multi_cohort_cgp_hist_plot(mapvizieR_obj, studentids, measurementscale,
  match_method = "no matching", first_and_spring_only = TRUE,
  entry_grade_seasons = c(-0.8, 4.2), school_norms = 2015,
  primary_cohort_only = TRUE, arrange_logic = c(ncol = length(plots)))
}
\arguments{
\item{mapvizieR_obj}{conforming mapvizieR obj}

\item{studentids}{vector of studentids}

\item{measurementscale}{target subject}

\item{match_method}{do we limit to matched students, and if so, how?
no matching = any student record in the studentids.
UNIMPLEMENTED METHODS / TODO
strict = only kids who appear in all terms
strict after imputation = impute first, then use stritc method
back one = look back one test term, and only include kids who can be matched}

\item{first_and_spring_only}{show all terms, or only entry & spring?
default is TRUE.}

\item{entry_grade_seasons}{which grade_level_seasons are entry grades?}

\item{school_norms}{c(2012, 2015).  what school norms to use?  default
is 2012.}

\item{primary_cohort_only}{will determine the most frequent cohort and limit to 
students in that cohort.  designed to handle discrepancies in grade/cohort
pattern caused by previous holdovers.  default is TRUE.}

\item{arrange_logic}{layout logic passed to grid.arrange.  default is
wide, one column per plot.  if, for instance, one row per plot was
desired, pass \code{c(nrow = length(plots))} to arrange_logic.}
}
\value{
a list of ggplotGrobs
}
\description{
see cohort_cgp_hist_plot for use.  Pass a vector of studentids
of *all* desired cohorts.  Plot will facet one plot per cohort.
}

