% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{munge_startdate}
\alias{munge_startdate}
\title{munge_startdate}
\usage{
munge_startdate(x)
}
\arguments{
\item{x}{teststartdate field}
}
\description{
helper function to convert a variety of date formats cleanly
}

