% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_wrappers.R
\name{mutate.mapvizieR_data}
\alias{mutate.mapvizieR_data}
\title{mutate wrapper}
\usage{
\method{mutate}{mapvizieR_data}(.data, ...)
}
\arguments{
\item{.data}{data.frame}

\item{...}{additional args}
}
\value{
data.frame
}
\description{
wrapper for mutate that preserves classes of data frames
}
