% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{mv_debugging_vars}
\alias{mv_debugging_vars}
\title{mv_debugging_vars}
\usage{
mv_debugging_vars()
}
\value{
injects some variables into your environment to ease function debugging
}
\description{
mv_debugging_vars
}

