% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/mv_filter.R
\name{mv_filter}
\alias{mv_filter}
\title{filter mapvizieR object}
\usage{
mv_filter(mapvizieR_obj, cdf_filter = NA, roster_filter = NA)
}
\arguments{
\item{mapvizieR_obj}{mapvizieR object}

\item{cdf_filter}{a filter, or filters to apply on fields in the cdf.  wrap it in \code{quote()}}

\item{roster_filter}{a filter, or filters to apply on fields in the roster.  will also filter
the cdf to only return those students.  wrap it in `quote()`}
}
\value{
a filtered mapvizieR object
}
\description{
filter a mapvizieR object by academic year, or any variable in the roster.
}

