% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{mv_limit_growth}
\alias{mv_limit_growth}
\title{mv_limit_growth_df}
\usage{
mv_limit_growth(mapvizieR_obj, studentids, measurementscale)
}
\arguments{
\item{mapvizieR_obj}{a conforming mapvizieR object}

\item{studentids}{vector of studentids}

\item{measurementscale}{a MAP subject}
}
\description{
extract the growth df and limit it to target students
}

