% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{mv_opening_checks}
\alias{mv_opening_checks}
\title{mv_opening_checks}
\usage{
mv_opening_checks(mapvizieR_obj, studentids, min_stu = 1)
}
\arguments{
\item{mapvizieR_obj}{a valid mapvizieR object.}

\item{studentids}{vector of studentids to run for this plot}

\item{min_stu}{minimum number of students for this plot.  default is 1.}
}
\description{
common beginning checks when building a plot.  DRY, right?
}

