% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/util.R
\name{n_timings}
\alias{n_timings}
\title{n_timings}
\usage{
n_timings(n, test_function, test_args)
}
\arguments{
\item{n}{num times to run the function}

\item{test_function}{name of the function, passed to do.call}

\item{test_args}{list of arguments for the function, passed
to do.call}
}
\description{
a convenience wrapper around timings to record
the results of timing a function's execution
}

