% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_RIT.R
\name{nearest_rit}
\alias{nearest_rit}
\title{find nearest RIT score for a student, by date}
\usage{
nearest_rit(mapvizieR_obj, studentid, measurementscale, target_date,
  num_days = 180, forward = TRUE)
}
\arguments{
\item{mapvizieR_obj}{mapvizieR object}

\item{studentid}{target studentid}

\item{measurementscale}{target subject}

\item{target_date}{date of interest, \code{Y-m-d} format}

\item{num_days}{function will only return test score within num_days of target_date}

\item{forward}{default is TRUE, set to FALSE if only scores before target_date should be chosen}
}
\description{
Given studentid, measurementscale, and a target_date, the function will return 
the closest RIT score.
}

