% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/norms_prep.R
\name{norms_students_wide_to_long}
\alias{norms_students_wide_to_long}
\title{norms_students_wide_to_long}
\usage{
norms_students_wide_to_long(norms = norms_students_2011)
}
\arguments{
\item{norms}{a norms_student object.  defaults to norms_students_2011}
}
\description{
takes the wide norms_students object and makes it long
}

