% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/college_plots.R
\name{npr_goal_sheet_style}
\alias{npr_goal_sheet_style}
\title{Wrapper to make student percentile rank goal sheet templates KIPP NJ style}
\usage{
npr_goal_sheet_style(measurementscale)
}
\arguments{
\item{measurementscale}{target subject}
}
\value{
a ggplot template
}
\description{
Wrapper to make student percentile rank goal sheet templates KIPP NJ style
}

