% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/cgp_prep.R
\name{npr_to_rit}
\alias{npr_to_rit}
\title{npr_to_rit}
\usage{
npr_to_rit(measurementscale, grade, season, npr)
}
\arguments{
\item{measurementscale}{MAP subject}

\item{grade}{grade level}

\item{season}{fall winter spring}

\item{npr}{a percentile rank, between 1-99}
}
\description{
given a percentile rank, return the best match RIT
}

