% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{numeric_nwea_seasons}
\alias{numeric_nwea_seasons}
\title{numeric_nwea_seasons}
\usage{
numeric_nwea_seasons(x)
}
\arguments{
\item{x}{a vector of seasons}
}
\value{
a vector of numeric offsets
}
\description{
tranforms fall, winter, spring into a numeric offset so things sort the right way
}
