% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mapvizieR_object.R
\name{nwea_growth}
\alias{nwea_growth}
\title{Create a data frame of all requested NWEA MAP growth norms.}
\usage{
nwea_growth(start.grade, start.rit, measurementscale, ...)
}
\arguments{
\item{start.grade}{vector of student start (or pre-test) grade levels}

\item{start.rit}{vector of student start (or pre-test) RIT scores}

\item{measurementscale}{vector of measurement scales for the RIT scores in \code{start.rit}}

\item{\code{...}}{arguments passed to dplyr:select, used to select the requested norms data.  You pass indicators like
T42, S22, R12 as unevaluated args (i.e. as unquoted strings).  For examples, passing R42 causes the
function to return a single vector of reported fall-to-spring growth norms; passing R42, S42, R22 would
return a data.frame with 3 columns for reported fall-to-spring growth, the standard deviation of fall-to-spring
growth and reported spring-spring-growth, respectively}
}
\value{
a vector of \code{length(start.grade)} or data.frame with \code{nrow(start.grade)} and \code{ncols(x)==length(...)}.
}
\description{
\code{nwea_growth} takes three vectors for grade-level, (starting) RIT Score, and Measurement Scale
 (usually from a CDF) and return a returns data.frame of typical growth
 calculations from the NWEA 2011 MAP Norms tables for each grade-RIT-measurement scale triplet.
}
\details{
User can indicate which calculted nomrs (typical mean, reported mean, and standard deviation) and
 any growth period by using the letter+two digit NWEA 2011 Growth Norms indicator (i.e. for
 Reported spring to spring growth user will provide R22, for typical fall to winter growth user provides
 T41, and for the the standard deviation of fall to spring growth user will provide S42). Providing no list
 of norms indicators results in every norm and season returned. All passed vectors must be the same length
}
\examples{
nwea_growth(start.grade = 5,
            start.rit = 190,
            measurementscale = "Reading")

# example with CDF
cdf<- ex_CombinedAssessmentResults \%>\%
        prep_cdf_long
roster <- prep_roster(ex_CombinedStudentsBySchool)
cdf$grade <-  grade_levelify_cdf(cdf, roster)

growth<-nwea_growth(start.grade = cdf$grade,
                    start.rit = cdf$testritscore,
                    measurementscale = cdf$measurementscale
)
cdf_2<-cbind(cdf, growth)

glimpse(cdf_2)

# get spring to spring growth stats only
growth_s2s<-nwea_growth(start.grade = cdf$grade,
                        start.rit = cdf$testritscore,
                        measurementscale = cdf$measurementscale,
                        'contains("22")'
)
cdf_3<-cbind(cdf, growth_s2s)
glimpse(cdf_3)
}

