% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linking.R
\name{ny_linking}
\alias{ny_linking}
\title{ny linking}
\usage{
ny_linking(measurementscale, current_grade, season, RIT,
  returns = "perf_level")
}
\arguments{
\item{measurementscale}{MAP subject}

\item{current_grade}{grade level}

\item{season}{c('Fall', 'Winter', 'Spring')}

\item{RIT}{rit score}

\item{returns}{one of c('perf_level', 'proficient')}
}
\value{
either character, length 1 with performance level or logical, 
length 1, with proficiency
}
\description{
ny linking
}
