% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_pass.R
\name{ny_pass_prob}
\alias{ny_pass_prob}
\title{calculate new york pass probability}
\usage{
ny_pass_prob(mapvizieR_obj)
}
\arguments{
\item{mapvizieR_obj}{conforming mapvizieR object}
}
\value{
cdf with predicted pass and predicted level
}
\description{
calculate new york pass probability
}
