% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linking_data.R
\docType{data}
\name{ny_predicted_pass}
\alias{ny_predicted_pass}
\title{New York predicted pass rate}
\format{\describe{
\item{grade}{student's grade level}
\item{measurementscale}{NWEA measurementscale}
\item{fallwinterspring}{season}
\item{percentile}{percentile, 2015 norms}
\item{pass_probability}{probability of achieving a level 3 or level 4}
}}
\source{
NWEA New York linking study
}
\usage{
ny_predicted_pass
}
\description{
Predicted NYS proficiency from NWEA linking study
}
\keyword{datasets}
