% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_data.R
\docType{data}
\name{ny_predicted_proficiency}
\alias{ny_predicted_proficiency}
\title{ny predicted proficiency}
\format{\describe{
\item{ny_subj}{ny_subj}
\item{ny_grade}{ny_grade}
\item{ny_season}{ny_season}
\item{ny_rit}{ny_rit}
\item{perf_level}{perf_level}
\item{proficient}{proficient}
}}
\source{
ny_proficiency.Rmd
}
\usage{
ny_predicted_proficiency
}
\description{
data from implied regression models in the NWEA nyc linking study
}
\keyword{datasets}

