% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{nys_math_3}
\alias{nys_math_3}
\alias{nys_math_4}
\alias{nys_read_3}
\alias{nys_read_4}
\title{ny state psuedo-quartiles}
\usage{
nys_math_3(x)

nys_math_4(x)

nys_read_3(x)

nys_read_4(x)
}
\arguments{
\item{x}{vector to compare (presumably percentile ranks)}
}
\value{
vector of 'quartiles'
}
\description{
ny state psuedo-quartiles
}

