% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cgp_prep.R
\name{one_cgp_step}
\alias{one_cgp_step}
\title{wrapper to simplify CGP simulations interface to simplify CGP calculations}
\usage{
one_cgp_step(measurementscale, start_rit, start_grade, cgp,
  growth_window = "Spring to Spring")
}
\arguments{
\item{measurementscale}{target subject}

\item{start_rit}{mean starting rit}

\item{start_grade}{starting grade (not grade level season, just grade)}

\item{cgp}{target cgp.  can be single target or vector}

\item{growth_window}{what growth window to step}
}
\description{
wrapper to simplify CGP simulations interface to simplify CGP calculations
}

