% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/report_dispatcher.R
\name{only_valid_plots}
\alias{only_valid_plots}
\title{only_valid_plots}
\usage{
only_valid_plots(x)
}
\arguments{
\item{x}{a list of report_dispatcher output}
}
\description{
a post-processor for report_dispatcher
}

