% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cgp_prep.R
\name{percentile_gain_needed}
\alias{percentile_gain_needed}
\title{percentile_gain_needed}
\usage{
percentile_gain_needed(target_cgp, grade_level, start_npr)
}
\arguments{
\item{grade_in,}{start_npr, target_cgp}
}
\description{
called by generalization method.
}

