% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf_prep.R
\name{prep_cdf_long}
\alias{prep_cdf_long}
\title{prep_cdf_long}
\usage{
prep_cdf_long(cdf_long)
}
\arguments{
\item{cdf_long}{a map assessmentresults.csv file.  can be one term, or many terms
together in one file.}
}
\value{
a prepped cdf file
}
\description{
\code{prep_cdf_long} a wrapper around several cdf prep functions
}
