% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/roster_prep.R
\name{prep_roster}
\alias{prep_roster}
\title{prep_roster}
\usage{
prep_roster(roster)
}
\arguments{
\item{students_by_school}{one, or multiple (combined) NWEA MAP
studentsbyschool.csv file(s).}
}
\value{
a prepped roster file
}
\description{
\code{prep_roster} a wrapper around several roster prep functions
}

