% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapvizieR_object.R
\name{print.mapvizieR}
\alias{print.mapvizieR}
\title{print method for \code{mapvizier} class}
\usage{
\method{print}{mapvizieR}(x, ...)
}
\arguments{
\item{x}{a \code{mapvizier} object}

\item{...}{additional arguments}
}
\value{
some details about the object to the console.
}
\description{
prints to console
}
\details{
Prints a summary of the a \code{mapvizier} object.
}
\examples{
\dontrun{
data(ex_CombinedAssessmentResults)
data(ex_CombinedStudentsBySchool)

cdf_mv <- mapvizieR(ex_CombinedAssessmentResults, 
                    ex_CombinedStudentsBySchool)
                    
cdf_mv
}
}

