% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf_prep.R
\name{process_cdf_long}
\alias{process_cdf_long}
\title{process_cdf_long}
\usage{
process_cdf_long(prepped_cdf)
}
\arguments{
\item{prepped_cdf}{output of prep_cdf_long}
}
\value{
a processed cdf file
}
\description{
\code{process_cdf_long} the second step in cdf processing
}

