% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cgp_prep.R
\name{project_cgp_targets}
\alias{project_cgp_targets}
\title{project_cgp_targets}
\usage{
project_cgp_targets(measurementscale, grade, growth_window, baseline_avg_rit,
  tolerance = 10, sch_growth_study = sch_growth_norms_2012,
  calc_for = c(1:99))
}
\arguments{
\item{measurementscale}{MAP subject}

\item{grade}{baseline/starting grad for the group of students}

\item{growth_window}{desired growth window for targets (fall/spring, spring/spring, fall/fall)}

\item{baseline_avg_rit}{the baseline mean rit for the group of students}

\item{tolerance}{NWEA has published empirical lookup tables for growth.  these tables cover
the middle of the distribution, meaning that data for most cohorts can be found in the table.
but say that you have a cohort that is 10 rit points below the lowest mean RIT reported in the
tables?  do you just use the lowest entry?
mapvizieR has a generalization of the school growth study that translate the RIT changes to
percentile change, and then fits a general model to express cohort growth percentile
given a starting percentile and an ending percentile. rather than use lookup tables that we
know to be poor fits to the students}

\item{school_growth_study}{a school growth study to use.  default is sch_growth_norms_2012}
}
\description{
given a baseline, what scores are necessary to reach certain growth targets
}

