% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quealy_subgroups.R
\name{quealy_facet_one_subgroup}
\alias{quealy_facet_one_subgroup}
\title{quealy_facet_one_subgroup}
\usage{
quealy_facet_one_subgroup(sum_df, subgroup, xlims, n_range, ref_lines = NA)
}
\arguments{
\item{sum_df}{output of quealy_permutation_stats.  needs to have a header
called facet_me.  look at quealy_subgroups to see example use.}

\item{subgroup}{the subgroup to plot.  quealy_subgroups calls this plot
once per element in subgroup_cols (and once for all students)}

\item{xlims}{the global xlims for the plot.}

\item{n_range}{the global range of n values.  used to set the width of
the lines}

\item{ref_lines}{if using with all students, the reference lines
showing change in all_students}
}
\value{
a data frame
}
\description{
the plot called by quealy subgroups for each subgroup.  used
to be internal to the function, has been extracted.
}

