% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quealy_subgroups.R
\name{quealy_permutation_stats}
\alias{quealy_permutation_stats}
\title{quealy_permutation_stats}
\usage{
quealy_permutation_stats(df, subgroup)
}
\arguments{
\item{df}{a growth data frame}

\item{subgroup}{the subgroup to group and calculate summary stats for}
}
\value{
a data frame
}
\description{
calculates group stats for all the permutations of a subroup.  used
to be internal to quealy_subgroups, has been extracted.
}

