% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/quealy_subgroups.R
\name{quealy_subgroups}
\alias{quealy_subgroups}
\title{quealy_subgroups}
\usage{
quealy_subgroups(mapvizieR_obj, studentids, measurementscale,
  subgroup_cols = c("starting_quartile"),
  pretty_names = c("Starting Quartile"), start_fws, start_academic_year,
  end_fws, end_academic_year, complete_obsv = FALSE)
}
\arguments{
\item{mapvizieR_obj}{mapvizieR object}

\item{studentids}{target students}

\item{measurementscale}{target subject}

\item{subgroup_cols}{what subgroups in mapvizier roster do you want to cut by?  default
is starting_quartile}

\item{pretty_names}{nicely formatted names for the column cuts used above.}

\item{start_fws}{starting season}

\item{start_academic_year}{starting academic year}

\item{end_fws}{ending season}

\item{end_academic_year}{ending academic year}

\item{complete_obsv}{if TRUE, limit only to students who have BOTH a start
and end score. default is FALSE.}
}
\description{
kevin quealy of the nytimes did a nice job capturing change
in the general population vs change in specific subgroups: http://nyti.ms/1tQrOIl
re: obamacare.
let's do the same thing for change in RIT score.
}

