% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/college_plots.R
\name{rainbow_colors}
\alias{rainbow_colors}
\title{creates vector of rainbow colors for student plots}
\usage{
rainbow_colors()
}
\value{
a vector of colors
}
\description{
creates vector of rainbow colors for student plots
}

