% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_util.R
\name{report_footer}
\alias{report_footer}
\title{Adds a simple footer to a report page at 1/20th the height of the existing
content}
\usage{
report_footer(page, footer_content)
}
\arguments{
\item{page}{probably the output of gridExtra::arrangeGrob}

\item{footer_content}{what you want the footer to say}
}
\value{
a grob with the footer content
}
\description{
Adds a simple footer to a report page at 1/20th the height of the existing
content
}

