% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/college_plots.R
\name{rit_height_weight_ACT}
\alias{rit_height_weight_ACT}
\title{Makes the base template that shows the RIT space and corresponding
ACT lines using the original NWEA MAP ACT linking study}
\usage{
rit_height_weight_ACT(measurementscale, color_list = rainbow_colors(),
  annotation_style = "points", line_style = "none",
  school_type = "MS", localization = localize("Newark"))
}
\arguments{
\item{measurementscale}{c('Reading', 'Mathematics') - there is no linking
study for Language or Science}

\item{color_list}{vector of colors to use to shade the bands.  Default is
the output of rainbow_colors().}

\item{annotation_style}{c('points', 'big numbers', or 'small numbers')}

\item{line_style}{c('gray lines', 'gray dashed')}

\item{school_type}{c('ES', 'MS')}

\item{localization}{controls names/breakpoints for college labels and
ACT tiers.  See localization.R for more details}
}
\value{
a ggplot object, to be used as a template for other plots
}
\description{
Makes the base template that shows the RIT space and corresponding
ACT lines using the original NWEA MAP ACT linking study
}
