% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/college_plots.R
\name{rit_height_weight_npr}
\alias{rit_height_weight_npr}
\title{Makes the base template that shows the RIT space and corresponding
percentile rank lines.}
\usage{
rit_height_weight_npr(measurementscale, color_list = rainbow_colors(),
  ribbon_alpha = 0.35, annotation_style = "points", line_style = "none",
  spring_only = TRUE, norms = 2015)
}
\arguments{
\item{measurementscale}{c('Reading', 'Mathematics') - there is no linking
study for Language or Science}

\item{color_list}{vector of colors to use to shade the bands.  Default is
the output of rainbow_colors().}

\item{ribbon_alpha}{transparency value for the background ribbons}

\item{annotation_style}{c('points', 'big numbers', or 'small numbers')}

\item{line_style}{c('gray lines')}

\item{spring_only}{fall norms show a 'summer slump' effect; this can be
visually distracting.  spring_only won't include those points in the reference
lines.}

\item{norms}{c(2011, 2015).  which norms study to use?}
}
\value{
a ggplot object, to be used as a template for other plots
}
\description{
Makes the base template that shows the RIT space and corresponding
percentile rank lines.
}

