% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cgp_prep.R
\name{rit_to_npr}
\alias{rit_to_npr}
\title{rit_to_npr}
\usage{
rit_to_npr(measurementscale, grade, season, RIT)
}
\arguments{
\item{measurementscale}{MAP subject}

\item{grade}{grade level}

\item{season}{fall winter spring}

\item{RIT}{rit score}
}
\description{
given a RIT score, return the best match percentile rank
}

