% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/roster_prep.R
\name{roster_prep_names}
\alias{roster_prep_names}
\title{roster_prep_names}
\usage{
roster_prep_names(students_by_school)
}
\arguments{
\item{students_by_school}{one, or multiple (combined) NWEA MAP
studentsbyschool.csv file(s).}
}
\value{
a roster with lowercase data frame names
}
\description{
\code{roster_prep_names} turns the CamelCase names of a StudentsBySchool to lowercase.
}

