% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/roster_to_df.R
\name{roster_to_cdf}
\alias{roster_to_cdf}
\title{roster_to_cdf}
\usage{
roster_to_cdf(target_df, mapvizieR_obj, roster_cols)
}
\arguments{
\item{target_df}{the df you want to put stuff on}

\item{mapvizieR_obj}{a conforming mapvizieR object}

\item{roster_cols}{roster column names you want to move over.}
}
\value{
a cdf data frame with the roster objects
}
\description{
when you need to put a roster object onto a cdf.  reasonably easy because this is
point in time data.
}

