% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/roster_to_df.R
\name{roster_to_growth_df}
\alias{roster_to_growth_df}
\title{roster_to_growth_df}
\usage{
roster_to_growth_df(target_df, mapvizieR_obj, roster_cols,
  disambiguation_method = "last")
}
\arguments{
\item{target_df}{the df you want to put stuff on}

\item{mapvizieR_obj}{a conforming mapvizieR object}

\item{roster_cols}{roster column names you want to move over.}

\item{disambiguation_method}{how to disambiguate?  default is 'last'.}
}
\value{
a growth data frame with the roster objects
}
\description{
when you need to put a roster object onto a growth data frame.
growth data frames are tricky, because they cover a time *span*, not a single point in time.
if a student changes school / grade / teacher (common) or if any of the other demographic attributes
of a student change (IEP status / lunch status / gender / etc.) you need to have a consistent rule for
attribution.  roster_to_growth_df implements that.
}

