% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{round_to_any}
\alias{round_to_any}
\title{round_to_any}
\usage{
round_to_any(x, accuracy, f = round)
}
\arguments{
\item{x}{numeric or date-time (POSIXct) vector to round}

\item{accuracy}{number to round to; for POSIXct objects, a number of seconds}

\item{f}{rounding function: \code{\link{floor}}, \code{\link{ceiling}} or
\code{\link{round}}

@return a numeric vector
@export}
}
\description{
because we don't want \code{suggests: plyr} if we can avoid it.
}

