% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mapvizieR_object.R
\name{s2s_match}
\alias{s2s_match}
\title{Merge two different assessment seasons (by student and measurement scale)
from a single long-format MAP assessment data frame}
\usage{
s2s_match(.data, season1 = "Fall", season2 = "Spring", sy = 2013,
  typical.growth = TRUE, college.ready = TRUE)
}
\arguments{
\item{.data}{a data frame with assessment data, in long-format (i.e., one student-assessment per row).}

\item{season1}{a string of either "Fall", "Winter", or "Spring" for the first assessment season by which to subset \code{.data} and join on.}

\item{season2}{a string of either "Fall", "Winter", or "Spring" for the first assessment season by which to subset \code{.data} and join on.
Note that if "Spring" to "Fall" is  a valid combination but norms cannot be calcualted for it.}

\item{sy}{an integer indicating the (second half school year.  For example, enter 2014 for the 2013-2104 school year. For assessment combinations that are ambiguious like
spring to spring, the school year is taken to be the the school year for season 2 and season 1 is taken from
the prior year.}

\item{typical.growth}{boolean indicating if typical growth, typical growth target, and typical growth met/exceeded inditor are to be calculted. Requires that \code{.data} have norm columns.}

\item{college.ready}{boolean indicating if college ready growth, college ready growth target, and college ready growth met/exceeded inditor are to be calculted. Requires that \code{.data} have KIPPTieredGrowth column.}
}
\value{
a data.frame ith at least 2(m-1) columns (and as many as $(m-1) + 6) and
a row for every student-assessment that occured in both season 1 and season 2.
}
\description{
\code{s2s_match} a dataframe with two season results matched on student-measurement
scale basis.
}
\details{
This function returns a data frame that results from subsetting a long-format data MAP assessment
data frame (i.e., where every students' test event occupies a single row), \code{.data} into two seasons
and then mergeing the two subsets with an inner join on student id and measurement
scale (via \code{dplyr::inner_join}) for the school year and seasons passed to it. All columns of \code{.data} are replicated, save of student id
and measuremen scale (since these are used to merge on) with the later season's column names
indicated wiht .2 suffix (i.e., \code{TestRITScore.2}).

If indicated (by setting the values of the  \code{typical.growth} and \code{college.ready}
parameters to \code{TRUE}) the function will also calculate the amount of growth (i.e. number of RIT points) and
growth target (i.e., RIT score to attain) college ready growth.

Note well that for \code{typical.growth=TRUE} the original
data frame, \code{.data}, must have reported norms for the requried growth season.  The reported norm column
must be named using the 2011 NWEA Norms table convention of the season (winter through fall) indicated by corresponding
integers (1-4);  For example fall to spring requires \code{.data} has a field names
\code{R42} and spring to spring wourld be \code{R22}.  These columns can be easily
added to a CDF by using the \code{\link{nwea_growth}} function.

Also note that calculating college ready growth and growth targets
requires that \code{.data} has a column named \code{KIPPTieredGrowth} containing
a KIPP tiered multiplier for each student-assessment.  These data can be generated
using the \code{\link{tiered_growth}} function.
}
\examples{
data(ex_CombinedAssessmentResults)
data(ex_CombinedStudentsBySchool)

require(dplyr)
cdf<- ex_CombinedAssessmentResults \%>\%
prep_cdf_long

roster <- prep_roster(ex_CombinedStudentsBySchool)

cdf$grade <-  grade_levelify_cdf(cdf, roster)

cdf_growth_ss<-s2s_match(cdf,
                         season1 = "Spring",
                         season2 = "Spring",
                         sy = 2013)
glimpse(cdf_growth_ss)
}

