% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/automated_cleaning.R
\name{sbs_missing}
\alias{sbs_missing}
\title{surfaces missing students by school data for a user creating a new mv object}
\usage{
sbs_missing(raw_sbs)
}
\arguments{
\item{raw_sbs}{}
}
\value{
nothing; simply prints out information about the students by school
}
\description{
surfaces missing students by school data for a user creating a new mv object
}
\examples{
\dontrun{
cdf<-read_cdf("data/")

str(cdf)
}
}

