% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mapvizier.R
\name{scale_color_quartile}
\alias{scale_color_quartile}
\title{Quartile Color Scale for ggplot2}
\usage{
scale_color_quartile(..., palette = "default")
}
\arguments{
\item{...}{Arguments passed to \code{\link[ggplot2]{scale_color_manual}}}

\item{palette}{Which palette to use: "default" or "kipp"}
}
\value{
A ggplot2 scale object
}
\description{
A convenience scale for applying consistent quartile colors to points and lines.
}
\examples{
library(ggplot2)

df <- data.frame(
  x = rnorm(100),
  y = rnorm(100),
  quartile = factor(sample(1:4, 100, replace = TRUE))
)
ggplot(df, aes(x = x, y = y, color = quartile)) +
  geom_point() +
  scale_color_quartile()

}
