% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mapvizier.R
\name{scale_fill_growth}
\alias{scale_fill_growth}
\title{Growth Status Fill Scale for ggplot2}
\usage{
scale_fill_growth(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[ggplot2]{scale_fill_manual}}}
}
\value{
A ggplot2 scale object
}
\description{
A convenience scale for applying consistent growth status colors.
}
