% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mapvizier.R
\name{scale_fill_quartile}
\alias{scale_fill_quartile}
\title{Quartile Fill Scale for ggplot2}
\usage{
scale_fill_quartile(..., palette = "default")
}
\arguments{
\item{...}{Arguments passed to \code{\link[ggplot2]{scale_fill_manual}}}

\item{palette}{Which palette to use: "default" or "kipp"}
}
\value{
A ggplot2 scale object
}
\description{
A convenience scale for applying consistent quartile colors to ggplot2 plots.
}
\examples{
library(ggplot2)

df <- data.frame(quartile = factor(1:4), count = c(25, 30, 28, 17))
ggplot(df, aes(x = quartile, y = count, fill = quartile)) +
  geom_col() +
  scale_fill_quartile()

}
