% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/school_growth_norms.R
\docType{data}
\name{sch_growth_norms_2012}
\alias{sch_growth_norms_2012}
\title{school growth norms 2012}
\format{\describe{
\item{grade}{grade}
\item{measurementscale}{measurementscale}
\item{growth_window}{growth_window}
\item{start_fallwinterspring}{start_fallwinterspring}
\item{end_fallwinterspring}{end_fallwinterspring}
\item{rit}{rit}
\item{typical_cohort_growth}{typical_cohort_growth}
\item{sd_of_expectation}{sd_of_expectation}
}}
\source{
http://legacysupport.nwea.org/sites/www.nwea.org/files/resources/2012%20RIT%20Scale%20School%20Norms%20User's%20Guide.pdf
}
\usage{
sch_growth_norms_2012
}
\description{
school growth norms 2012
}
\keyword{datasets}

