% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/schambach_table.R
\name{schambach_table}
\alias{schambach_table}
\title{schambach_table}
\usage{
schambach_table(mapvizieR_obj, measurementscale_is, grade,
  subgroup_cols = c("end_schoolname"), pretty_names = c("School Name"),
  start_fws, start_academic_year, end_fws, end_academic_year,
  complete_obsv = FALSE)
}
\arguments{
\item{mapvizieR_obj}{mapvizieR object}

\item{measurementscale_is}{target subject}

\item{grade}{target grade}

\item{subgroup_cols}{what subgroups, in addition to end_schoolname, in mapvizier roster do you want to cut by?}

\item{pretty_names}{nicely formatted names for the column cuts used above.}

\item{start_fws}{starting season}

\item{start_academic_year}{starting academic year}

\item{end_fws}{ending season}

\item{end_academic_year}{ending academic year}

\item{complete_obsv}{if TRUE, limit only to students who have BOTH a start
and end score. default is FALSE.}
}
\description{
use similar methods in quealy_subgroups to create tables based on tables
Lindsay Schambach is making. initial cut will be on schools
}

