% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/schambach_table.R
\name{schambach_table_1d}
\alias{schambach_table_1d}
\title{schambach_table}
\usage{
schambach_table_1d(mapvizieR_obj, measurementscale_in, studentids,
  subgroup_cols = c("grade"), pretty_names = c("Grade"), start_fws,
  start_academic_year, end_fws, end_academic_year, complete_obsv = FALSE)
}
\arguments{
\item{mapvizieR_obj}{mapvizieR object}

\item{measurementscale_in}{target subject}

\item{studentids}{target studentids}

\item{subgroup_cols}{what subgroups to explore, default is by grade}

\item{pretty_names}{nicely formatted names for the column cuts used above.}

\item{start_fws}{starting season}

\item{start_academic_year}{starting academic year}

\item{end_fws}{ending season}

\item{end_academic_year}{ending academic year}

\item{complete_obsv}{if TRUE, limit only to students who have BOTH a start
and end score. default is FALSE.}
}
\description{
\code{schambach_table} Given grade level(s), shows summary table for provided subgroups.
Table includes average ending RIT, percent started in top 75%, percented ended in 75%, average
percentile growth, percent meeting "Keep Up", percent meeting "Rutgers Ready", and the number of
students in each subgroup. Named after Lindsay Schambach, who provided initial table.
}

