% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/norm_data.R
\docType{data}
\name{school_status_norms_2015_dense_extended}
\alias{school_status_norms_2015_dense_extended}
\title{school_status_norms_2015_dense_extended 2015 norm data densified and extended}
\format{\describe{
\item{measurementscale}{measurementscale}
\item{fallwinterspring}{fallwinterspring}
\item{grade}{grade}
\item{RIT}{RIT}
\item{student_percentile}{student_percentile}
\item{school_percentile}{school_percentile}
}}
\source{
http://support.nwea.org/support/article/rit-scale-norms-study-data-files
}
\usage{
school_status_norms_2015_dense_extended
}
\description{
Builds on status_norms_2015 for schools.
Two data processing steps have been performed:
1) densification by percentile.
2) extending norms into HS.
See student_status_norms_2015_dense_extended for more details
}
\keyword{datasets}

