% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_data.R
\docType{data}
\name{school_status_norms_2025_dense_extended}
\alias{school_status_norms_2025_dense_extended}
\title{school_status_norms_2025_dense_extended 2025 school norm data densified and extended}
\format{
\describe{
\item{measurementscale}{measurementscale}
\item{fallwinterspring}{fallwinterspring}
\item{grade}{grade}
\item{RIT}{RIT}
\item{school_percentile}{school_percentile}
}
}
\source{
NWEA 2025 MAP Growth Technical Manual
}
\usage{
school_status_norms_2025_dense_extended
}
\description{
Builds on status_norms_2025 for schools.
Two data processing steps have been performed:
1) densification by percentile
2) extending norms into HS grades
See student_status_norms_2025_dense_extended for more details.
Note: Original 2025 norms only include 5-point percentile intervals.
}
\keyword{datasets}
