% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/growth_df_prep.R
\name{scores_by_testid}
\alias{scores_by_testid}
\title{scores_by_testid}
\usage{
scores_by_testid(testid, processed_cdf, start_or_end)
}
\arguments{
\item{testid}{a vector of testids}

\item{processed_cdf}{a conforming processed_cdf data frame}

\item{start_or_end}{either c('start', 'end')}
}
\value{
one row data frame.
}
\description{
helper function for \code{generate_growth_df}. given a test id,
returns df with all the scores.
}

