% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_rit.R
\name{season_fix}
\alias{season_fix}
\title{season fix}
\usage{
season_fix(cdf)
}
\arguments{
\item{cdf}{a cdf, after imputation}
}
\value{
the cdf, with a repaired fallwinterspring column
}
\description{
fixes the season column of a cdf after imputation
}
