% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_wrappers.R
\name{select.mapvizieR_data}
\alias{select.mapvizieR_data}
\title{select wrapper}
\usage{
select.mapvizieR_data(.data, ...)
}
\arguments{
\item{.data}{data.frame}

\item{...}{additional args}
}
\value{
data.frame
}
\description{
wrapper for select that preserves classes of data frames
}
