% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sgp_histogram.R
\name{sgp_histogram}
\alias{sgp_histogram}
\title{sgp_histogram}
\usage{
sgp_histogram(mapvizieR_obj, studentids, measurementscale, start_fws,
  start_academic_year, end_fws, end_academic_year, perf_breaks = c(55, 45))
}
\arguments{
\item{mapvizieR_obj}{mapvizieR object}

\item{studentids}{target students}

\item{measurementscale}{target subject}

\item{start_fws}{starting season}

\item{start_academic_year}{starting academic year}

\item{end_fws}{ending season}

\item{end_academic_year}{ending academic year}

\item{perf_breaks}{sgp_histogram will color the median growth percentile
green, yellow, or red.  where to break between colors?  default is 55 and 45.}
}
\value{
returns a ggplot object
}
\description{
\code{sgp_histogram} a simple visualization of the distribution of student SGPs.
}

