% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzz_test.R
\name{silly_plot}
\alias{silly_plot}
\title{silly plot}
\usage{
silly_plot(mapvizieR_obj, studentids)
}
\arguments{
\item{mapvizieR_obj}{a conforming mapvizieR object}

\item{studentids}{a vector of studentids}
}
\description{
just a dead simple plot that we can use to test the fuzz_test function
}

