% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{standardize_kinder}
\alias{standardize_kinder}
\title{Standardize Kindergarten code to 0 and other grades to integers}
\usage{
standardize_kinder(x, other_codes = NULL)
}
\arguments{
\item{x}{a character, integer, or numeric vector of grade levels}

\item{other_codes}{a vector of alternative codes for kindergarten to be translated to 0}
}
\value{
an integer vector of \code{length(x)}.
}
\description{
\code{standardize_kinder} returns an integer vector grade levels.
}
\details{
This function simply translates kindergarten codes 13 (from client-based
MAP), "K", or a user specified \code{kinder_code} from an input vector \code{x}
 to 0 and recasts the  input vector to an integer vector.
}
\examples{
# Create vector of grades 1 through 13, where 13 = kinder
x <- sample(x=1:13, 100,replace = TRUE)
standardize_kinder(x)

# change 13 to "K"
x2 <- ifelse(x==13, "K", x)
standardize_kinder(x2)

# change "K" to "Kinder"
x3 <- ifelse(x=="K", "Kinder", x)
standardize_kinder(x2, other_codes="Kinder")
}

