% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_pass.R
\name{state_pass_prob}
\alias{state_pass_prob}
\title{state pass probability}
\usage{
state_pass_prob(mapvizieR_obj, state_test = "NY")
}
\arguments{
\item{mapvizieR_obj}{a valid mapvizieR object}

\item{state_test}{name of the state test to calculate for}
}
\value{
a cdf with predicted pass and predicted level
}
\description{
state pass probability
}
