% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/norm_data.R
\docType{data}
\name{status_norms_2015}
\alias{status_norms_2015}
\title{status_norms_2015 NWEA student and school status (or, attainment)
norms data (2015 study)}
\format{\describe{
\item{measurementscale}{measurementscale}
\item{fallwinterspring}{fallwinterspring}
\item{grade}{grade}
\item{RIT}{RIT}
\item{student_percentile}{student_percentile}
\item{school_percentile}{school_percentile}
}}
\source{
http://support.nwea.org/support/article/rit-scale-norms-study-data-files

http://support.nwea.org/support/article/rit-scale-norms-study-data-files
}
\usage{
status_norms_2015

status_norms_2015
}
\description{
Norm data published by NWEA:
Given a subject, grade level, season and RIT, gives a percentile rank

Norm data published by NWEA:
Given a subject, grade level, pre-test and post-test seasons and start RIT, expected growth
and variance statistics.
}
\keyword{datasets}

