% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_data.R
\docType{data}
\name{status_norms_2020}
\alias{status_norms_2020}
\title{status_norms_2020 NWEA student and school status norms data (2020 study)}
\format{
\describe{
\item{measurementscale}{measurementscale}
\item{fallwinterspring}{fallwinterspring}
\item{grade}{grade}
\item{RIT}{RIT}
\item{student_percentile}{student_percentile}
\item{school_percentile}{school_percentile}
}
}
\source{
NWEA 2020 MAP Growth Normative Data
}
\usage{
status_norms_2020
}
\description{
Norm data published by NWEA in the 2020 MAP Growth Normative Data report.
Given a subject, grade level, season and RIT, gives a percentile rank.
}
\keyword{datasets}
