% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_data.R
\docType{data}
\name{status_norms_2025}
\alias{status_norms_2025}
\title{status_norms_2025 NWEA student and school status norms data (2025 study)}
\format{
\describe{
\item{measurementscale}{measurementscale}
\item{fallwinterspring}{fallwinterspring}
\item{grade}{grade}
\item{RIT}{RIT}
\item{student_percentile}{student_percentile}
\item{school_percentile}{school_percentile}
}
}
\source{
NWEA 2025 MAP Growth Technical Manual
}
\usage{
status_norms_2025
}
\description{
Norm data published by NWEA in the 2025 MAP Growth Technical Manual.
Given a subject, grade level, season and RIT, gives a percentile rank.
Note: 2025 norms only include percentiles at 5-point intervals (5, 10, 15, ..., 95).
}
\keyword{datasets}
