% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/strand_boxes.R
\name{strand_boxes}
\alias{strand_boxes}
\title{goal strand boxes}
\usage{
strand_boxes(mapvizieR_obj, studentids, measurementscale_in, fws, academic_year)
}
\arguments{
\item{mapvizieR_obj}{mapvizieR object}

\item{studentids}{target students}

\item{measurementscale_in}{target subject}

\item{start_fws}{starting season}

\item{start_academic_year}{starting academic year}

\item{end_fws}{ending season}

\item{end_academic_year}{ending academic year}
}
\value{
prints a ggplot object
}
\description{
goal strand boxes
}

