% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strand_list_plot.R
\name{strands_list_plot}
\alias{strands_list_plot}
\title{Creates lists of students names stacked by rank in each of a given 
MAP assessments goal strands}
\usage{
strands_list_plot(mapvizier_obj, studentids, measurement_scale, season, year)
}
\arguments{
\item{mapvizier_obj}{a \code{\link{mapvizieR}} object}

\item{studentids}{a set of student ids to subset to}

\item{measurement_scale}{the subject to subest to}

\item{season}{the season to use}

\item{year}{academic year}
}
\value{
a ggplot object
}
\description{
Creates lists of students names stacked by rank in each of a given 
MAP assessments goal strands
}
\examples{
\dontrun{
require(dplyr)

data("ex_CombinedStudentsBySchool")
data("ex_CombinedAssessmentResults")

map_mv <- mapvizieR(ex_CombinedAssessmentResults, ex_CombinedStudentsBySchool)

ids <- ex_CombinedStudentsBySchool \%>\% 
   dplyr::filter(
     Grade == 8,
     SchoolName == "Mt. Bachelor Middle School",
     TermName == "Spring 2013-2014") \%>\% 
   dplyr::select(StudentID) \%>\%
   unique()

strands_list_plot(map_mv, 
                 ids$StudentID, 
                 "Mathematics", 
                 "Fall", 
                 2013)
}
}

