% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_detail_table.R
\name{stu_growth_detail}
\alias{stu_growth_detail}
\title{Student growth detail}
\usage{
stu_growth_detail(mapvizieR_obj, studentids, measurementscale,
  entry_grade_seasons = c(-0.8, 4.2))
}
\arguments{
\item{mapvizieR_obj}{a conforming mapvizieR object}

\item{studentids}{vector of studentids}

\item{measurementscale}{target subject}

\item{entry_grade_seasons}{what grades are considered entry grades}
}
\value{
a data frame with growth data
}
\description{
calculates longitudinal / summary growth stats for many students,
allowing analysis like 'who grew the most'? and 'who had the largest average 
conditional growth percentile'?
}
