% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_detail_table.R
\name{stu_growth_detail_table}
\alias{stu_growth_detail_table}
\title{Student Growth Detail Table}
\usage{
stu_growth_detail_table(mapvizieR_obj, studentids, measurementscale,
  high_or_low_growth = "high", num_stu = 5,
  entry_grade_seasons = c(-0.8, 4.2), title = "", ...)
}
\arguments{
\item{mapvizieR_obj}{a conforming mapvizieR object}

\item{studentids}{vector of studentids}

\item{measurementscale}{target subject}

\item{high_or_low_growth}{should we list most, or least growth?
default is high.}

\item{num_stu}{how many students to show in the table? default is 5.}

\item{entry_grade_seasons}{what grades are considered entry grades}

\item{title}{if desired, a title for the table}

\item{...}{additional arguments to tableGrob}
}
\value{
a tableGrob
}
\description{
growth detail in printable format.  wrapper around 
student_growth_detail
}
