% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/norm_data.R
\docType{data}
\name{student_growth_norms_2011}
\alias{student_growth_norms_2011}
\title{student_growth_norms_2011 NWEA student growth
norms data (2011 study)}
\format{\describe{
\item{Subject}{Measurement Scale, Possible Values: 1 = Mathematics, 2 = Reading,
3 = Language Usage, 4 = General science}
\item{StartGrade}{The RIT score of the test in the first term of the comparison period.
Scores are now represented with as many as 14‐decimal places, which is clearly overboard.}
\item{T41}{Growth projection for comparison period fall of this grade to winter of this grade.
Negative numbers likely due to extrapolation.}
\item{T42}{Growth projection for comparison period fall of this grade to spring of this grade.
Negative numbers likely due to extrapolation}
\item{T44}{Growth projection for comparison period fall of this grade to fall of next grade.
Negative numbers likely due to extrapolation}
\item{T22}{Growth projection for comparison period spring of this grade to spring of next grade.
Negative numbers likely due to extrapolation}
\item{T12}{Growth projection for comparison period winter of this grade to spring of this grade
Negative numbers likely due to extrapolation}
\item{R41}{Reported growth projection for comparison period fall of this grade to winter of this grade}
\item{R42}{Reported growth projection for comparison period fall of this grade to spring of this grade}
\item{R44}{Reported growth projection for comparison period fall of this grade to fall of next grade}
\item{R22}{Reported growth projection for comparison period spring of last grade to spring of this grade}
\item{R12}{Reported growth projection for comparison period winter of this grade to spring of this grade}
\item{S41}{Standard deviation of growth projection for comparison period fall of this grade to winter of this grade}
\item{S42}{Standard deviation growth projection for comparison period fall of this grade to spring of this grade}
\item{S44}{Standard deviation growth projection for comparison period fall of this grade to fall of next grade}
\item{S22}{Standard deviation growth projection for comparison period spring of last grade to spring of this grade}
\item{S12}{Standard deviation growth projection for comparison period winter of this grade to spring of this grade}
\item{MeasurementScale}{Measurement scale in plain English}
\item{norms_year}{year norms were published, in this case 2011}
}}
\source{
http://support.nwea.org/support/article/rit-scale-norms-study-data-files
}
\usage{
student_growth_norms_2011
}
\description{
Norm data published by NWEA:
Given a subject, grade level, pre-test and post-test seasons and start RIT, expected growth
and variance statistics.
}
\keyword{datasets}

