% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/student_npr_history_plot.R
\name{student_npr_history_plot}
\alias{student_npr_history_plot}
\title{A small multiples plot of a set of students' national percentile rank histories.}
\usage{
student_npr_history_plot(mv_obj, studentids, measurment_scale)
}
\arguments{
\item{mv_obj}{a \code{\link{mapvizier}} object}

\item{studentids}{a set of student ids to subset too.}

\item{measurment_scale}{the subject of interest}
}
\value{
a ggplot2 object
}
\description{
A small multiples plot of a set of students' national percentile rank histories.
}
\examples{
data("ex_CombinedStudentsBySchool")
data("ex_CombinedAssessmentResults")

map_mv<-mapvizieR(ex_CombinedAssessmentResults, ex_CombinedStudentsBySchool)

ids<-ex_CombinedStudentsBySchool \%>\% filter(Grade==8, SchoolName=="Mt. Bachelor Middle School", TermName=="Spring 2013-2014") \%>\% select(StudentID) \%>\% unique()

student_npr_history_plot(map_mv, studentids = ids[1:80, "StudentID"], measurment_scale = "Reading")
}

