% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student_npr_two_term_plot.R
\name{student_npr_two_term_plot}
\alias{student_npr_two_term_plot}
\title{A small multiples plot of a set of students' national percentile rank over two terms.}
\usage{
student_npr_two_term_plot(mapvizieR_obj, studentids,
  measurement_scale = "Reading", term_first, term_second, n_col = 9,
  min_n = 10)
}
\arguments{
\item{mapvizieR_obj}{a \code{\link{mapvizieR}} object}

\item{studentids}{a set of student ids to subset to}

\item{measurement_scale}{a MAP measurementscale}

\item{term_first}{the first test term. You need to use the full name, such
as 'Spring 2012-2013'}

\item{term_second}{the second test term. You need to use the full name, such
as 'Fall 2013-2014'}

\item{n_col}{number of columns to display in the plot}

\item{min_n}{minimum number of students required to produce the plot.}
}
\value{
a ggplot object
}
\description{
A small multiples plot of a set of students' national percentile rank over two terms.
}
\examples{
\dontrun{
require(dplyr)

data("ex_CombinedStudentsBySchool")
data("ex_CombinedAssessmentResults")

map_mv <- mapvizieR(ex_CombinedAssessmentResults, ex_CombinedStudentsBySchool)

ids <- ex_CombinedStudentsBySchool \%>\% 
   dplyr::filter(
     Grade == 8,
     SchoolName == "Mt. Bachelor Middle School",
     TermName == "Spring 2013-2014") \%>\% 
   dplyr::select(StudentID) \%>\%
   unique()

student_npr_two_term_plot(
  map_mv,
  studentids = ids[1:80, "StudentID"],
  measurement_scale ="Reading", 
  term_first = "Spring 2012-2013", 
  term_second = "Fall 2013-2014", 
  n_col = 7, 
  min_n = 5)
}
}

