% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/norm_data.R
\docType{data}
\name{student_status_norms_2011}
\alias{student_status_norms_2011}
\title{student_status_norms_2011 NWEA student norm data (2011 study)}
\format{\describe{
\item{measurementscale}{measurementscale}
\item{fallwinterspring}{fallwinterspring}
\item{grade}{grade}
\item{RIT}{RIT}
\item{percentile}{percentile}
}}
\source{
http://support.nwea.org/support/article/rit-scale-norms-study-data-files
}
\usage{
student_status_norms_2011
}
\description{
Norm data published by NWEA:
Given a subject, grade level, season and RIT, gives a percentile rank
}
\keyword{datasets}

