% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_data.R
\docType{data}
\name{student_status_norms_2020_dense_extended}
\alias{student_status_norms_2020_dense_extended}
\title{student_status_norms_2020_dense_extended 2020 student norm data densified and extended}
\format{
\describe{
\item{measurementscale}{measurementscale}
\item{fallwinterspring}{fallwinterspring}
\item{grade}{grade}
\item{RIT}{RIT}
\item{student_percentile}{student_percentile}
}
}
\source{
NWEA 2020 MAP Growth Normative Data
}
\usage{
student_status_norms_2020_dense_extended
}
\description{
Builds on status_norms_2020 for students.
Two data processing steps have been performed:
1) densification by percentile - creates entries for intermediate percentiles
2) extending norms into HS grades where data may be sparse
}
\keyword{datasets}
