% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_data.R
\docType{data}
\name{student_status_norms_2025_dense_extended}
\alias{student_status_norms_2025_dense_extended}
\title{student_status_norms_2025_dense_extended 2025 student norm data densified and extended}
\format{
\describe{
\item{measurementscale}{measurementscale}
\item{fallwinterspring}{fallwinterspring}
\item{grade}{grade}
\item{RIT}{RIT}
\item{student_percentile}{student_percentile}
}
}
\source{
NWEA 2025 MAP Growth Technical Manual
}
\usage{
student_status_norms_2025_dense_extended
}
\description{
Builds on status_norms_2025 for students.
Two data processing steps have been performed:
1) densification by percentile - creates entries for intermediate percentiles
2) extending norms into HS grades where data may be sparse
Note: Original 2025 norms only include 5-point percentile intervals.
}
\keyword{datasets}
