% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_wrappers.R
\name{summarize.mapvizieR_data}
\alias{summarize.mapvizieR_data}
\title{summarize wrapper}
\usage{
\method{summarize}{mapvizieR_data}(.data, ..., .by = NULL, .groups = NULL)
}
\arguments{
\item{.data}{data.frame}

\item{...}{additional args}

\item{.by}{see \code{\link[dplyr]{summarize}}}

\item{.groups}{see \code{\link[dplyr]{summarize}}}
}
\value{
data.frame
}
\description{
wrapper for summarize that preserves classes of data frames
}
