% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapvizier_summary.R
\name{summary.mapvizieR_cdf}
\alias{summary.mapvizieR_cdf}
\title{summary method for \code{mapvizieR_cdf} class}
\usage{
\method{summary}{mapvizieR_cdf}(object, ...)
}
\arguments{
\item{object}{a \code{mapvizieR_cdf} object}

\item{...}{other arguments to be passed to other functions (not currently supported)}
}
\value{
summary stats as a \code{mapvizier_cdf_summary} object.
}
\description{
summary method for \code{mapvizieR_cdf} class
}
