% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapvizier_summary.R
\name{summary.mapvizieR_growth}
\alias{summary.mapvizieR_growth}
\title{summary method for \code{mapvizieR_growth} class}
\usage{
\method{summary}{mapvizieR_growth}(object, ...)
}
\arguments{
\item{object}{a \code{mapvizieR_growth} object}

\item{...}{other arguments to be passed to other functions (not currently supported)}
}
\value{
summary stats as a \code{mapvizier_summary} object.
}
\description{
summarizes growth data from \code{mapvizieR_growth} object.
}
\details{
Creates a \code{mapvizier_growth_summary} object of growth data from a \code{mapvizieR} 
object.  Includes the following summarizations for every growth term available
in the \code{mapvizier_growth} object:
\itemize{
 \item number tested in both assessment seasons (i.e., the number of students who 
 too a test in both assessment season and for which we are able to calcualate growth stats).
 \item Total students making typical growth
 \item Percent of students making typical growth
 \item Total students making college ready growth
 \item Percent of students making college ready  growth
 \item Total students with NPR >= 50 percentile in the first assessment season
 \item Percent students with NPR >= 50 percentile in the first assessment season
 \item Total students with NPR >= 75th percentile in the first assessment season
 \item Percent students with NPR >= 75 percentile in the first assessment season
 \item Total students with NPR >= 50 percentile in the second assessment season
 \item Percent students with NPR >= 50 percentile in the second assessment season
 \item Total students with NPR >= 75th percentile in the second assessment season
 \item Percent students with NPR >= 75 percentile in the second assessment season
}
}

