% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/teacher_performance_report.R
\name{teacher_performance_update}
\alias{teacher_performance_update}
\title{Teacher Performance Update}
\usage{
teacher_performance_update(mapvizieR_obj, studentids, measurementscale,
  start_fws, start_academic_year, end_fws, end_academic_year, title_text = "",
  ...)
}
\arguments{
\item{mapvizieR_obj}{mapvizieR object}

\item{studentids}{target students}

\item{measurementscale}{target subject}

\item{start_fws}{starting season}

\item{start_academic_year}{starting academic year}

\item{end_fws}{ending season}

\item{end_academic_year}{ending academic year}

\item{title_text}{what is this report called?}

\item{...}{additional arguments}

\item{detail_academic_year}{don't mask any data for this academic year}
}
\value{
prints a ggplot object
}
\description{
Teacher Performance Update
}

