% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{template_01}
\alias{template_01}
\title{template 1}
\usage{
template_01(p01, p02)
}
\arguments{
\item{p01}{plot or grob}

\item{p02}{plot or grob}
}
\value{
grob, output of arrangeGrob
}
\description{
(1 x 1) columns
}
