% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{template_02}
\alias{template_02}
\title{template 2}
\usage{
template_02(p01, p02, p03)
}
\arguments{
\item{p01}{plot or grob}

\item{p02}{plot or grob}

\item{p03}{plot or grob}
}
\value{
grob, output of arrangeGrob
}
\description{
(1 x 1 x 1)
}

