% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{template_03}
\alias{template_03}
\title{template 3}
\usage{
template_03(p01, p02, p03, p04)
}
\arguments{
\item{p01}{plot or grob}

\item{p02}{plot or grob}

\item{p03}{plot or grob}

\item{p04}{plot or grob}
}
\value{
grob, output of arrangeGrob
}
\description{
row 1: (1 x 1), row 2 (1 x 1)
}
