% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{template_05}
\alias{template_05}
\title{template 5}
\usage{
template_05(p01, p02)
}
\arguments{
\item{p01}{plot or grob}

\item{p02}{plot or grob}
}
\value{
grob, output of arrangeGrob
}
\description{
(1 x 1), rows
}

