% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{template_06}
\alias{template_06}
\title{template 6}
\usage{
template_06(p01, p02, p03, p04, p05, p06)
}
\arguments{
\item{p01}{plot or grob}

\item{p02}{plot or grob}

\item{p03}{plot or grob}

\item{p04}{plot or grob}

\item{p05}{plot or grob}

\item{p06}{plot or grob}
}
\value{
grob, output of arrangeGrob
}
\description{
col 1: (1 x 1), col 2: 1, col 3: (1 x 1)
}
