% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/testing_util.R
\name{testing_constants}
\alias{testing_constants}
\title{testing_constants}
\usage{
testing_constants()
}
\description{
wrapper function that calls populate_constants if they aren't in the
environment
}

