% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{tg}
\alias{tg}
\title{tg "to grob" helper function}
\usage{
tg(x)
}
\arguments{
\item{x}{a report object. currently supports grobs or ggplot objects}
}
\value{
a grob, suitable for arrangeGrob on a report layout
}
\description{
tg "to grob" helper function
}
