% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{tiered_growth_factors}
\alias{tiered_growth_factors}
\title{Calculate KIPP Tiered Growth factors}
\usage{
tiered_growth_factors(quartile, grade)
}
\arguments{
\item{quartile}{a vector of student quartiles}

\item{grade}{vector of student grade-levels}
}
\value{
a vector of \code{length(quartile)} of KIPP Tiered Growth factors.
}
\description{
\code{tiered_growth_factors} takes grade level and quartile data and returns a vector of KIPP 
Tiered Growth factors (or multipliers, if you prefer).
}
\details{
# Function takes two vectors---one containing student grade levels and the other 
 containing student pre-test/season 1 quartiles---and returns a same-length vector of 
 KIPP Tired Growth factors.  These factors are multiplied by a students typical 
 (i.e., expected) growth to generate college ready growth.
}
