% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/util.R
\name{time_execution}
\alias{time_execution}
\title{time_execution}
\usage{
time_execution(n, test_function, test_args)
}
\arguments{
\item{n}{num times to run the function}

\item{test_function}{name of the function, passed to do.call}

\item{test_args}{list of arguments for the function, passed
to do.call}
}
\description{
times how long it takes to execute a function
}

