% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_wrappers.R
\name{ungroup_.mapvizieR_data}
\alias{ungroup_.mapvizieR_data}
\title{ungroup wrapper}
\usage{
ungroup_.mapvizieR_data(df, ...)
}
\arguments{
\item{df}{data.frame}

\item{...}{additional args}
}
\value{
data.frame
}
\description{
wrapper for ungroup that preserves classes of data frames
}

