% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/util.R
\name{valid_grade_seasons}
\alias{valid_grade_seasons}
\title{valid_grade_seasons}
\usage{
valid_grade_seasons(cdf, entry_grade_seasons = c(-0.8, 4.2),
  first_and_spring_only = TRUE, detail_academic_year = 2014)
}
\arguments{
\item{cdf}{a processed cdf}

\item{entry_grade_seasons}{which grade seasons are 'entry' for this school?}
}
\description{
a filter on a cdf that restricts the grade_level_season ONLY to
spring data, and fall of 'entry' grades
}

