% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_wrappers.R
\name{wrapper_class_orderer}
\alias{wrapper_class_orderer}
\title{class orderer for dplyr wrappers}
\usage{
wrapper_class_orderer(all_classes)
}
\arguments{
\item{all_classes}{character vector of class names}
}
\value{
character vector of classes with anything `mapviz` sorted first
}
\description{
class orderer for dplyr wrappers
}

