% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_wrappers.R
\name{arrange.mapvizieR_data}
\alias{arrange.mapvizieR_data}
\title{arrange wrapper}
\usage{
\method{arrange}{mapvizieR_data}(df, ...)
}
\arguments{
\item{df}{data.frame}

\item{...}{additional args}
}
\value{
data.frame
}
\description{
wrapper for arrange that preserves classes of data frames
}
