% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgp_prep.R
\name{calc_cgp}
\alias{calc_cgp}
\title{calc_cgp}
\usage{
calc_cgp(measurementscale, end_grade, growth_window, baseline_avg_rit = NA,
  ending_avg_rit = NA, norms = 2015, calc_for = c(1:99), verbose = TRUE)
}
\arguments{
\item{measurementscale}{MAP subject}

\item{end_grade}{baseline/starting grad for the group of students}

\item{growth_window}{desired growth window for targets (fall/spring, spring/spring, fall/fall)}

\item{baseline_avg_rit}{the baseline mean rit for the group of students}

\item{ending_avg_rit}{the baseline mean rit for the group of students}

\item{norms}{which school growth study to use.  c(2012, 2015).  default is
2015.}

\item{calc_for}{vector of cgp targets to calculate for.}

\item{verbose}{should warnings about invalid seasons be raised?}
}
\value{
a named list - targets, and results
}
\description{
calculates both cgp targets and cgp results.
}

