% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgp_prep.R
\name{cdf_to_cgp}
\alias{cdf_to_cgp}
\title{CDF to CGP summary}
\usage{
cdf_to_cgp(cdf, grouping = "implicit_cohort")
}
\arguments{
\item{cdf}{conforming cdf file}

\item{grouping}{what column to group on.  default is implicit_cohort}
}
\value{
a data frame with summary start/end rit, and cgps
}
\description{
CDF to CGP summary
}

