% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cohort_cgp_hist.R
\name{cohort_cgp_hist_plot}
\alias{cohort_cgp_hist_plot}
\title{Show's a cohort's progress over time, in percentile space.}
\usage{
cohort_cgp_hist_plot(mapvizieR_obj, studentids, measurementscale,
  first_and_spring_only = TRUE, school_norms = 2015)
}
\arguments{
\item{mapvizieR_obj}{conforming mapvizieR obj}

\item{studentids}{vector of studentids}

\item{measurementscale}{target subject}

\item{first_and_spring_only}{show all terms, or only entry & spring?
default is TRUE.}

\item{school_norms}{c(2012, 2015).  what school norms to use?  default
is 2012.}
}
\value{
a ggplot object
}
\description{
Show's a cohort's progress over time, in percentile space.
}

