% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cgp_prep.R
\name{determine_cgp_method}
\alias{determine_cgp_method}
\title{determine_cgp_method}
\usage{
determine_cgp_method(measurementscale_in, grade_in, growth_window_in,
  baseline_avg_rit, tolerance, sch_growth_study)
}
\arguments{
\item{grade_in}{baseline/starting grad for the group of students}

\item{growth_window_in}{desired growth window for targets (fall/spring, spring/spring, fall/fall)}

\item{baseline_avg_rit}{the baseline mean rit for the group of students}

\item{tolerance}{threshold to depart from lookup to generalization}

\item{sch_growth_study}{which NWEA school growth study to use}

\item{measurementscale}{MAP subject}
}
\description{
determine_cgp_method
}

