% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goals_and_projections.R
\name{ensure_goals_names}
\alias{ensure_goals_names}
\title{ensure_goals_names}
\usage{
ensure_goals_names(.)
}
\arguments{
\item{.}{dot-placeholder, per ensurer doc.}
}
\description{
a contract that ensures that a goal object's has the 
proper elements.
}

