% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/util.R
\name{ensure_nonzero_students_with_norms}
\alias{ensure_nonzero_students_with_norms}
\title{ensure_nonzero_students_with_norms}
\usage{
ensure_nonzero_students_with_norms(.)
}
\arguments{
\item{.}{dot-placeholder, per ensurer doc.}
}
\description{
a contract that verifies that a growth df has at least one student with
normative data
}

