% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/roster_check.R
\name{ensure_roster_types}
\alias{ensure_roster_types}
\title{ensure_roster_types}
\usage{
ensure_roster_types(.)
}
\arguments{
\item{.}{dot-placeholder, per ensurer doc.}
}
\description{
checks if certain roster fields are of the correct type
}

