% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/util.R
\name{ensure_rows_in_df}
\alias{ensure_rows_in_df}
\title{ensure_rows_in_df}
\usage{
ensure_rows_in_df(.)
}
\arguments{
\item{.}{dot-placeholder, per ensurer doc.}
}
\description{
a contract that verifies that a data set isn't length zero
}

