% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fall_goals_report.R
\name{fall_goals_report}
\alias{fall_goals_report}
\title{Fall Goals Report}
\usage{
fall_goals_report(mapvizieR_obj, studentids, measurementscale, context,
  start_fws = "Spring", start_year_offset = -1, end_fws = "Spring",
  end_academic_year = 2015, entry_grade_seasons = c(-0.8, 4.2),
  detail_academic_year = 9999, goal_cgp = 80, school_type = "MS")
}
\arguments{
\item{mapvizieR_obj}{a valid mapvizieR_object}

\item{studentids}{a vector of studentids to run}

\item{measurementscale}{desired subject}

\item{context}{what school/grade/class/etc grouping is represented?}

\item{start_fws}{character, starting season for school growth norms}

\item{start_year_offset}{0 if start season is same, -1 if start is prior year.}

\item{end_fws}{ending season}

\item{end_academic_year}{ending academic year}

\item{entry_grade_seasons}{for becca plot.  default is c(-0.8, 4.2)}

\item{detail_academic_year}{passed through to various plots}

\item{goal_cgp}{what target CGP should be used for goals?}

\item{school_type}{c('ES', 'MS')}
}
\value{
a multipage report, represented as a list of grobs.
}
\description{
Fall Goals Report
}

