% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/util.R
\name{fall_spring_me}
\alias{fall_spring_me}
\title{fall_spring_me}
\usage{
fall_spring_me(x)
}
\arguments{
\item{x}{a grade level between -1 and 12}
}
\value{
a labeled string
}
\description{
\code{fall_spring_me} tranforms grade levels into labels for charts; eg 4.2 -> F5
}

