% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_wrappers.R
\name{filter.mapvizieR_data}
\alias{filter.mapvizieR_data}
\title{filter wrapper}
\usage{
\method{filter}{mapvizieR_data}(.data, ..., .by = NULL, .preserve = FALSE)
}
\arguments{
\item{.data}{data.frame}

\item{...}{additional args}

\item{.by}{see \code{\link[dplyr]{filter}}}

\item{.preserve}{see \code{\link[dplyr]{filter}}}
}
\value{
data.frame
}
\description{
wrapper for filter that preserves classes of data frames
}
