% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/goalbar.R
\name{goalbar}
\alias{goalbar}
\title{goal_bar}
\usage{
goalbar(mapvizieR_obj, studentids, measurementscale, start_fws,
  start_academic_year, end_fws, end_academic_year,
  goal_labels = c(accel_growth = "Made Accel Growth", typ_growth =
  "Made Typ Growth", positive_but_not_typ = "Below Typ Growth", negative_growth
  = "Negative Growth", no_start = sprintf("Untested: \%s \%s", start_fws,
  start_academic_year), no_end = sprintf("Untested: \%s \%s", end_fws,
  end_academic_year)), goal_colors = c("#CC00FFFF", "#0066FFFF", "#CCFF00FF",
  "#FF0000FF", "#FFFFFF", "#F0FFFF"), ontrack_prorater = NA,
  ontrack_fws = NA, ontrack_academic_year = NA,
  ontrack_labels = c(ontrack_accel = "On Track for Accel Growth", ontrack_typ
  = "On Track for Typ Growth", offtrack_typ = "Off Track for Typ Growth"),
  ontrack_colors = c("#CC00FFFF", "#0066FFFF", "#CCFF00FF"),
  complete_obsv = FALSE)
}
\arguments{
\item{mapvizieR_obj}{mapvizieR object}

\item{studentids}{target students}

\item{measurementscale}{target subject}

\item{start_fws}{starting season}

\item{start_academic_year}{starting academic year}

\item{end_fws}{ending season}

\item{end_academic_year}{ending academic year}

\item{goal_labels}{what labels to show for each goal category.  must be in order from
highest to lowest.}

\item{goal_colors}{what colors to show for each goal category}

\item{ontrack_prorater}{default is NA.  if set to a decimal value, what percent of the goal
is considered ontrack?}

\item{ontrack_fws}{season to use for determining ontrack status}

\item{ontrack_academic_year}{year to use for determining ontrack status}

\item{ontrack_labels}{what labels to use for the 3 ontrack statuses}

\item{ontrack_colors}{what colors to use for the 3 ontrack colors}

\item{complete_obsv}{if TRUE, limit only to students who have BOTH a start
and end score. default is FALSE.}
}
\description{
a simple bar chart that shows the percentage of a cohort at different goal
states (met / didn't meet)
}

