% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_wrappers.R
\name{group_by.mapvizieR_data}
\alias{group_by.mapvizieR_data}
\title{group_by wrapper}
\usage{
group_by.mapvizieR_data(
  .data,
  ...,
  .add = FALSE,
  .drop = dplyr::group_by_drop_default(.data)
)
}
\arguments{
\item{.data}{data.frame}

\item{...}{additional args}

\item{.add}{see \code{\link[dplyr]{group_by}}}

\item{.drop}{see \code{\link[dplyr]{group_by}}}
}
\value{
data.frame
}
\description{
wrapper for group_by that preserves classes of data frames
}
