% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_status.R
\name{growth_status_scatter}
\alias{growth_status_scatter}
\title{growth status scatter}
\usage{
growth_status_scatter(
  mapvizieR_obj,
  studentids,
  measurementscale,
  start_fws,
  start_academic_year,
  end_fws,
  end_academic_year
)
}
\arguments{
\item{mapvizieR_obj}{mapvizieR object}

\item{studentids}{target students}

\item{measurementscale}{target subject}

\item{start_fws}{starting season}

\item{start_academic_year}{starting academic year}

\item{end_fws}{ending season}

\item{end_academic_year}{ending academic year}
}
\value{
a ggplot object
}
\description{
a scatterplot comparing student growth percentile, and attainment
(or status) percentile
}
