% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf_prep.R
\name{make_npr_consistent}
\alias{make_npr_consistent}
\title{make_npr_consistent}
\usage{
make_npr_consistent(cdf, norm_study = 2015)
}
\arguments{
\item{cdf}{a mostly-processed cdf object (this is the last step)
in process_cdf}

\item{norm_study}{c(2011, 2015).  year a norm study.  default is 2015.  look in norm_data.R
for documentation of available norm studies.}
}
\description{
join a cdf to a norms study and get the empirical 
percentiles.  protects against longitudinal findings being 
clouded by changes in the norms.
}

