% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgp_prep.R
\name{mapviz_cgp_targets}
\alias{mapviz_cgp_targets}
\title{mapvizieR interface to simplify growth goal calculations}
\usage{
mapviz_cgp_targets(mapvizieR_obj, studentids, measurementscale, start_fws,
  start_year_offset, end_fws, end_academic_year, end_grade,
  start_fws_prefer = NA, calc_for = c(1:99), returns = "targets")
}
\arguments{
\item{mapvizieR_obj}{mapvizieR object}

\item{studentids}{target students}

\item{measurementscale}{target subject}

\item{start_fws}{one academic season (if known); 
pass vector of two and mapviz_cgp_targets will pick}

\item{start_year_offset}{0 if start season is same, -1 if start is prior year.}

\item{end_fws}{ending season}

\item{end_academic_year}{ending academic year}

\item{end_grade}{specify the ending grade for the growth (this can't be reliably
inferred from data).}

\item{start_fws_prefer}{which term is preferred? not required if only one start_fws is passed}

\item{calc_for}{passed through to calc_cgp, what values to calculate targets for?}

\item{returns}{'targets' or 'expectations'?}
}
\value{
data frame of cgp targets
}
\description{
given an explicit window, or a composite baseline, will calculate
CGP targets
}

