% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mapvizieR_object.R
\name{mapvizieR}
\alias{mapvizieR}
\title{Create a mapvizieR object}
\usage{
mapvizieR(cdf, roster, verbose = FALSE, ...)
}
\arguments{
\item{cdf}{a NWEA AssessmentResults.csv or CDF}

\item{roster}{a NWEA students}

\item{verbose}{should mapvizieR print status updates?  default is FALSE.}
}
\description{
\code{mapvizieR} is a workhorse workflow function that
calls a sequence of cdf and roster prep functions, given a cdf and roster
}
\examples{
cdf_mv <- mapvizieR(ex_CombinedAssessmentResults,
                    ex_CombinedStudentsBySchool)

is.mapvizieR(cdf_mv)
}

