% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mapvizieR_object.R
\name{mapvizieR}
\alias{mapvizieR}
\title{Create a mapvizieR object}
\usage{
mapvizieR(raw_cdf, raw_roster)
}
\arguments{
\item{raw_cdf}{a NWEA AssessmentResults.csv or CDF}

\item{raw_roster}{a NWEA students}
}
\description{
\code{mapvizieR} is a workhorse workflow function that
calls a sequence of cdf and roster prep functions, given a raw cdf and raw roster
}
\examples{
cdf_mv <- mapvizieR(ex_CombinedAssessmentResults,
                    ex_CombinedStudentsBySchool)

is.mapvizieR(cdf_mv)
}

