% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapvizieR-package.R, R/mapvizieR_object.R
\docType{package}
\name{mapvizieR}
\alias{mapvizieR}
\alias{mapvizieR-package}
\title{mapvizieR}
\usage{
mapvizieR(cdf, roster, verbose = FALSE, norms = 2015, ...)
}
\arguments{
\item{cdf}{a NWEA AssessmentResults.csv or CDF}

\item{roster}{a NWEA students}

\item{verbose}{should mapvizieR print status updates?  default is FALSE.}

\item{norms}{norm study to use. passed through to cdf prep}

\item{...}{additional arguments to pass to constructor functions called by mapvizieR}
}
\description{
mapvizieR

\code{mapvizieR} is a workhorse workflow function that
calls a sequence of cdf and roster prep functions, given a cdf and roster
}
\examples{
\dontrun{
cdf_mv <- mapvizieR(ex_CombinedAssessmentResults, 
                    ex_CombinedStudentsBySchool)
                    
is.mapvizieR(cdf_mv)                     
}
}

