% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_wrappers.R
\name{mutate_.mapvizieR_data}
\alias{mutate_.mapvizieR_data}
\title{mutate wrapper}
\usage{
\method{mutate_}{mapvizieR_data}(df, ...)
}
\arguments{
\item{df}{data.frame}

\item{...}{additional args}
}
\value{
data.frame
}
\description{
wrapper for mutate that preserves classes of data frames
}
