% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{mv_limit_cdf}
\alias{mv_limit_cdf}
\title{mv_limit_cdf}
\usage{
mv_limit_cdf(mapvizieR_obj, studentids, measurementscale)
}
\arguments{
\item{mapvizieR_obj}{a conforming mapvizieR object}

\item{studentids}{vector of studentids}

\item{measurementscale}{a MAP subject}
}
\value{
the cdf
}
\description{
extract the cdf and limit it to target students
}
