% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgp_prep.R
\name{npr_to_rit}
\alias{npr_to_rit}
\title{npr_to_rit}
\usage{
npr_to_rit(measurementscale, current_grade, season, npr, norms = 2015)
}
\arguments{
\item{measurementscale}{MAP subject}

\item{current_grade}{grade level}

\item{season}{fall winter spring}

\item{npr}{a percentile rank, between 1-99}

\item{norms}{which norm study to use}
}
\value{
a integer vector length one
}
\description{
given a percentile rank, return the best match RIT
(assumes the subject is a student, not a school/cohort.)
}

