% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nyt_subgroups.R
\name{nyt_subgroups}
\alias{nyt_subgroups}
\title{nyt_subgroups}
\usage{
nyt_subgroups(mapvizieR_obj, studentids, measurementscale, start_fws,
  start_academic_year, end_fws, end_academic_year, complete_obsv = FALSE)
}
\arguments{
\item{mapvizieR_obj}{mapvizieR object}

\item{studentids}{target students}

\item{measurementscale}{target subject}

\item{start_fws}{starting season}

\item{start_academic_year}{starting academic year}

\item{end_fws}{ending season}

\item{end_academic_year}{ending academic year}

\item{complete_obsv}{if TRUE, limit only to students who have BOTH a start
and end score. default is FALSE.}
}
\description{
the times did a nice job capturing change in the population vs
change in subgroups: http://nyti.ms/1tQrOIl
let's do the same thing for change in RIT score.
}

