% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cgp_prep.R
\name{percentile_gain_needed}
\alias{percentile_gain_needed}
\title{percentile_gain_needed}
\usage{
percentile_gain_needed(target_cgp, grade_level, start_npr)
}
\arguments{
\item{target_cgp}{the CGP target}

\item{grade_level}{grade level (end of window)}

\item{start_npr}{avg start percentile rank}
}
\description{
given a CGP target, tells you how much improvement in class percentile
rank neeeded
}

