% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{quartile_order}
\alias{quartile_order}
\title{quartile_order}
\usage{
quartile_order(x)
}
\arguments{
\item{x}{a quartile (1-4)}
}
\description{
helper function used by becca plot to put quartiles in correct order
}

