% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quealy_subgroups.R
\name{quealy_subgroups}
\alias{quealy_subgroups}
\title{quealy_subgroups}
\usage{
quealy_subgroups(mapvizieR_obj, studentids, measurementscale,
  subgroup_cols = c("starting_quartile"),
  pretty_names = c("Starting Quartile"), start_fws, start_year_offset,
  end_fws, end_academic_year, start_fws_prefer = NA, report_title = NA,
  complete_obsv = FALSE, drop_NA_groups = TRUE, include_all = TRUE)
}
\arguments{
\item{mapvizieR_obj}{mapvizieR object}

\item{studentids}{target students}

\item{measurementscale}{target subject}

\item{subgroup_cols}{what subgroups in mapvizier roster do you want to cut by?  default
is starting_quartile}

\item{pretty_names}{nicely formatted names for the column cuts used above.}

\item{start_fws}{one academic season (if known); pass vector of two and quealy subgroups will pick}

\item{start_year_offset}{0 if start season is same, -1 if start is prior year.}

\item{end_fws}{ending season}

\item{end_academic_year}{ending academic year}

\item{start_fws_prefer}{which term is preferred? not required if only one start_fws is passed}

\item{report_title}{text grob to put on the report tile}

\item{complete_obsv}{if TRUE, limit only to students who have BOTH a start
and end score. default is FALSE.}

\item{drop_NA_groups}{should we ignore subgroups with value NA?  default is true}

\item{include_all}{should the output have plot at the top showing the TOTAL
variation?  not recommended for data spanning multiple grade levels.}
}
\value{
a grob composed of multiple ggplots
}
\description{
kevin quealy of the nytimes did a nice job capturing change
in the general population vs change in specific subgroups: http://nyti.ms/1tQrOIl
re: obamacare.
let's do the same thing for change in RIT score.
}

