% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{rand_stu}
\alias{rand_stu}
\title{rand_stu}
\usage{
rand_stu(mapvizieR_obj, low = 20, high = 500)
}
\arguments{
\item{mapvizieR_obj}{conforming mapvizieR object}

\item{low}{how many kids? lower bound}

\item{high}{how many kids? higher bound}
}
\description{
gives back a random sample of studentids from a mapvizieR object
}

