% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roster_to_df.R
\name{roster_to_cdf}
\alias{roster_to_cdf}
\title{roster_to_cdf}
\usage{
roster_to_cdf(target_df, mapvizieR_obj, roster_cols,
  by_measurementscale = FALSE)
}
\arguments{
\item{target_df}{the df you want to put stuff on}

\item{mapvizieR_obj}{a conforming mapvizieR object}

\item{roster_cols}{roster column names you want to move over.}

\item{by_measurementscale}{boolean, when you have student demographics that are specific to a 
particular assessment - eg course enrollment, but the match is specific to student AND
measurementscale, not just student.  if TRUE your roster object must contain a field
called measurementscale.}
}
\value{
a cdf data frame with the roster objects
}
\description{
when you need to put a roster object onto a cdf.  reasonably easy because this is 
point in time data.
}

