% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/cgp_prep.R
\name{sch_growth_lookup}
\alias{sch_growth_lookup}
\title{sch_growth_lookup}
\usage{
sch_growth_lookup(measurementscale_in, grade_in, growth_window_in,
  baseline_avg_rit, sch_growth_study = sch_growth_norms_2012)
}
\arguments{
\item{measurementscale_in}{MAP subject}

\item{grade_in}{baseline/starting grade for the group of students}

\item{growth_window_in}{desired growth window for targets (fall/spring, spring/spring, fall/fall)}

\item{baseline_avg_rit}{the baseline mean rit for the group of students}

\item{sch_growth_study}{NWEA school growth study to use for lookup; defaults to 2012.}
}
\description{
get cohort growth expectations via lookup from growth study
}

