% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cgp_prep.R
\name{sch_growth_lookup}
\alias{sch_growth_lookup}
\title{sch_growth_lookup}
\usage{
sch_growth_lookup(measurementscale, grade, growth_window, baseline_avg_rit,
  sch_growth_study = sch_growth_norms_2012)
}
\arguments{
\item{measurementscale}{MAP subject}

\item{growth_window}{desired growth window for targets (fall/spring, spring/spring, fall/fall)}

\item{baseline_avg_rit}{the baseline mean rit for the group of students}

\item{sch_growth_study}{NWEA school growth study to use for lookup; defaults to 2012.}

\item{grade_in}{baseline/starting grade for the group of students}
}
\description{
get cohort growth expectations via lookup from growth study
}

