% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgp_prep.R
\name{sch_growth_lookup}
\alias{sch_growth_lookup}
\title{sch_growth_lookup}
\usage{
sch_growth_lookup(
  measurementscale,
  end_grade,
  growth_window,
  baseline_avg_rit,
  norms = 2015
)
}
\arguments{
\item{measurementscale}{MAP subject}

\item{end_grade}{grade students will be in at the end of the window}

\item{growth_window}{desired growth window for targets (fall/spring, spring/spring, fall/fall)}

\item{baseline_avg_rit}{the baseline mean rit for the group of students}

\item{norms}{which school growth study to use.  c(2012, 2015).  default is
2015.}
}
\description{
get cohort growth expectations via lookup from growth study
}
