% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/norm_data.R, R/school_growth_norms.R
\docType{data}
\name{sch_growth_norms_2012}
\alias{sch_growth_norms_2012}
\title{school growth norms 2012 study}
\format{\describe{
\item{grade}{END grade, eg grade 2 spring to spring is grade 1 TO grade 2}
\item{measurementscale}{measurementscale}
\item{growth_window}{growth_window}
\item{start_fallwinterspring}{start_fallwinterspring}
\item{end_fallwinterspring}{end_fallwinterspring}
\item{typical_cohort_growth}{typical_cohort_growth}
\item{sd_of_expectation}{sd_of_expectation}
}}
\source{
NWEA 2012 school growth study (PDF)

http://legacysupport.nwea.org/sites/www.nwea.org/files/resources/2012%20RIT%20Scale%20School%20Norms%20User's%20Guide.pdf
}
\usage{
sch_growth_norms_2012

sch_growth_norms_2012
}
\description{
school / cohort growth norms published by NWEA
these are scraped out of the published PDF tables

school growth norms 2012
}
\keyword{datasets}

