% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/schambach_figure.R
\name{schambach_figure}
\alias{schambach_figure}
\title{schambach_fig}
\usage{
schambach_figure(mapvizieR_obj, measurementscale_in, studentids_in,
  subgroup_cols = c("grade"), pretty_names = c("Grade"), start_fws,
  start_academic_year, end_fws, end_academic_year, report_title = NA,
  complete_obsv = FALSE)
}
\arguments{
\item{mapvizieR_obj}{mapvizieR object}

\item{measurementscale_in}{target subject}

\item{studentids_in}{target studentids}

\item{subgroup_cols}{what subgroups to explore, default is by grade}

\item{pretty_names}{nicely formatted names for the column cuts used above.}

\item{start_fws}{starting season}

\item{start_academic_year}{starting academic year}

\item{end_fws}{ending season}

\item{end_academic_year}{ending academic year}

\item{complete_obsv}{if TRUE, limit only to students who have BOTH a start
and end score. default is FALSE.}

\item{report}{title text grob to put on report}
}
\description{
Produces list of figures from list of data.frames returned by schambach_table_1d
}

