% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_wrappers.R
\name{select_.mapvizieR_data}
\alias{select_.mapvizieR_data}
\title{select wrapper}
\usage{
\method{select_}{mapvizieR_data}(df, ...)
}
\arguments{
\item{df}{data.frame}

\item{...}{additional args}
}
\value{
data.frame
}
\description{
wrapper for select that preserves classes of data frames
}
