% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_data.R
\docType{data}
\name{status_norms_2015}
\alias{status_norms_2015}
\title{status_norms_2015 NWEA student and school status (or, attainment) 
norms data (2015 study)}
\format{
\describe{
\item{measurementscale}{measurementscale}
\item{fallwinterspring}{fallwinterspring}
\item{grade}{grade}
\item{RIT}{RIT}
\item{student_percentile}{student_percentile}
\item{school_percentile}{school_percentile}
}
}
\source{
http://support.nwea.org/support/article/rit-scale-norms-study-data-files
}
\usage{
status_norms_2015
}
\description{
Norm data published by NWEA: 
Given a subject, grade level, season and RIT, gives a percentile rank
}
\keyword{datasets}
