% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strand_boxes.R
\name{strand_boxes}
\alias{strand_boxes}
\title{goal strand boxes}
\usage{
strand_boxes(mapvizieR_obj, studentids, measurementscale, fws, academic_year)
}
\arguments{
\item{mapvizieR_obj}{mapvizieR object}

\item{studentids}{target students}

\item{measurementscale}{target subject}

\item{fws}{season}

\item{academic_year}{academic year}
}
\value{
a ggplot object
}
\description{
a plot that shows multiple goal scores as boxplots
}

