% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/student_npr_history_plot.R
\name{student_npr_history_plot}
\alias{student_npr_history_plot}
\title{A small multiples plot of a set of students' national percentile rank histories.}
\usage{
student_npr_history_plot(mapvizieR_obj, studentids, subject, title_text = "")
}
\arguments{
\item{mapvizieR_obj}{a \code{\link{mapvizieR}} object}

\item{studentids}{a set of student ids to subset too.}

\item{subject}{the subject of interest}

\item{title_text}{title as a character vector}
}
\value{
a ggplot2 object
}
\description{
This plots a proto-typical example of Edward Tufte's small multiples
concept.  A class of students in one subject and grade is plotted, witch each
student's historical national percentile rank plotted with a simple liear fit.

The background colors indicate the quartile the assessment earned.
}
\examples{
\dontrun{
require(dplyr)

data("ex_CombinedStudentsBySchool")
data("ex_CombinedAssessmentResults")

map_mv<-mapvizieR(ex_CombinedAssessmentResults, ex_CombinedStudentsBySchool)

ids<-ex_CombinedStudentsBySchool \%>\% filter(
   Grade==8,
   SchoolName=="Mt. Bachelor Middle School",
   TermName=="Spring 2013-2014") \%>\% select(StudentID) \%>\%
   unique()

student_npr_history_plot(
    map_mv,
    studentids = ids[1:80, "StudentID"],
    subject = "Reading")
}
}

