% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_wrappers.R
\name{summarize_.mapvizieR_data}
\alias{summarize_.mapvizieR_data}
\title{summarize wrapper}
\usage{
summarize_.mapvizieR_data(df, ...)
}
\arguments{
\item{df}{data.frame}

\item{...}{additional args}
}
\value{
data.frame
}
\description{
wrapper for summarize that preserves classes of data frames
}
