% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapvizier_summary.R
\name{summary.mapvizieR}
\alias{summary.mapvizieR}
\title{summary method for \code{mapvizieR} class}
\usage{
\method{summary}{mapvizieR}(object, ...)
}
\arguments{
\item{object}{a \code{mapvizieR} object}

\item{...}{other arguments to be passed to other functions (not currently supported)}
}
\value{
summary stats as a \code{mapvizier_summary} object.
}
\description{
produces a summary for all of the objects on the
main mapvizieR object.  specifically returns \code{mapvizieR_growth_summary}
and \code{mapvizieR_cdf_summary}
}

