% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/two_pager.R
\name{two_pager}
\alias{two_pager}
\title{MAP two-pager}
\usage{
two_pager(mapvizieR_obj, studentids, measurementscale, start_fws,
  start_academic_year, end_fws, end_academic_year, detail_academic_year,
  title_text = "", ...)
}
\arguments{
\item{mapvizieR_obj}{mapvizieR object}

\item{studentids}{target students}

\item{measurementscale}{target subject}

\item{start_fws}{starting season}

\item{start_academic_year}{starting academic year}

\item{end_fws}{ending season}

\item{end_academic_year}{ending academic year}

\item{detail_academic_year}{don't mask any data for this academic year}

\item{title_text}{what is this report called?}

\item{...}{additional arguments}
}
\value{
prints a ggplot object
}
\description{
MAP two-pager
}

