% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_wrappers.R
\name{ungroup.mapvizieR_data}
\alias{ungroup.mapvizieR_data}
\title{ungroup wrapper}
\usage{
ungroup.mapvizieR_data(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{additional args}
}
\value{
data.frame
}
\description{
wrapper for ungroup that preserves classes of data frames
}
